### 2022-05-13 #4391 @editorialbot
[PRE REVIEW]: RELSAD: A Python package for reliability assessment of modern distribution systems
**Submitting author:** @stinefm (<a href="http://orcid.org/0000-0002-2283-1724">Stine Fleischer Myhre</a>)
**Repository:** https://github.com/stinefm/relsad
**Branch with paper.md** (empty if default branch):
**Version:** v0.0.1
