+++ README.rst
@@ -1,0 +1,57 @@
+#################################################
+RELSAD
+#################################################
+
+![](https://github.com/stinefm/relsad/workflows/build/badge.svg) [![codecov](https://codecov.io/gh/stinefm/relsad/branch/main/graph/badge.svg)](https://codecov.io/gh/stinefm/relsad) ![Release](https://github.com/stinefm/relsad/workflows/Release/badge.svg) [![Documentation Status](https://readthedocs.org/projects/relsad/badge/?version=latest)](https://relsad.readthedocs.io/en/latest/?badge=latest)
+
+A Python package for reliability assessment of modern distribution systems.
+
+.....................................
+Installation
+.....................................
+
+Install in developer mode using:
+
+.. code-block:: console
+
+    python setup.py develop
+
+
+.. code-block:: console
+
+    $ pip install -i https://test.pypi.org/simple/ relsad
+
+.....................................
+Features
+.....................................
+
+- Reliability analysis
+- Load flow
+- Monte Carlo simulation
+
+.....................................
+Dependencies
+.....................................
+
+- Numpy
+- Scipy
+- Matplotlib
+- Pandas
+
+.....................................
+Usage
+.....................................
+
+See :ref:`usage`.
+
+.....................................
+Documentation
+.....................................
+
+The official documentation is hosted on Read the Docs: https://relsad.readthedocs.io/en/latest/
+
+.....................................
+Contributors
+.....................................
+
+We welcome and recognize all contributions. You can see a list of current contributors in the [contributors tab](https://github.com/stinefm/relsad/graphs/contributors).
