+++ README.md
@@ -1,43 +1,0 @@
-# The relsad package
-
-
-![](https://github.com/stinefm/relsad/workflows/build/badge.svg) [![codecov](https://codecov.io/gh/stinefm/relsad/branch/main/graph/badge.svg)](https://codecov.io/gh/stinefm/relsad) ![Release](https://github.com/stinefm/relsad/workflows/Release/badge.svg) [![Documentation Status](https://readthedocs.org/projects/relsad/badge/?version=latest)](https://relsad.readthedocs.io/en/latest/?badge=latest)
-
-This cookiecutter creates a boilerplate for a Python project.
-
-## Installation
-
-Install in developer mode using:
-
-```
-python setup.py develop
-```
-
-```bash
-$ pip install -i https://test.pypi.org/simple/ relsad
-```
-
-## Features
-
-- TODO
-
-## Dependencies
-
-- TODO
-
-## Usage
-
-- TODO
-
-## Documentation
-
-The official documentation is hosted on Read the Docs: https://relsad.readthedocs.io/en/latest/
-
-## Contributors
-
-We welcome and recognize all contributions. You can see a list of current contributors in the [contributors tab](https://github.com/stinefm/relsad/graphs/contributors).
-
-### Credits
-
-This package was created with Cookiecutter and the UBC-MDS/cookiecutter-ubc-mds project template, modified from the [pyOpenSci/cookiecutter-pyopensci](https://github.com/pyOpenSci/cookiecutter-pyopensci) project template and the [audreyr/cookiecutter-pypackage](https://github.com/audreyr/cookiecutter-pypackage).
-
