+++ docs/requirements.txt
@@ -1,0 +1,2 @@
+sphinxcontrib-bibtex = "^2.4.2"
+sphinx-book-theme = "^0.3.2"
