### 2022-05-14 14f24d7
fixed bug in readthedocs config
