+++ docs/source/index.rst
@@ -1,17 +1,6 @@
 Welcome to relsad's documentation!
 ===========================================================
 
-.. image:: https://img.shields.io/badge/tutorial-py--pkgs-blue
-        :target: https://ubc-mds.github.io/py-pkgs/
-
-.. image:: https://img.shields.io/badge/cookiecutter-ubc--mds-yellow
-        :target: https://github.com/UBC-MDS/cookiecutter-ubc-mds
-
-.. image:: https://img.shields.io/badge/badges-shields.io-brightgreen
-        :target: https://shields.io/
-
-This cookiecutter creates a boilerplate for a Python project.
-
 To get started, check out the sections below:
 
 .. toctree::
