+++ .readthedocs.yaml
@@ -5,7 +5,7 @@ version: 2
 
 sphinx:
   builder: html
-  configuration: docs/conf.py
+  configuration: docs/source/conf.py
 
 # Build all the formats
 formats:
