+++ relsad/test_networks/RBTS6.py
@@ -4,12 +4,11 @@ from relsad.network.components import (
     CircuitBreaker,
     Disconnector,
     Line,
-    ManualMainController, 
-
+    ManualMainController,
 )
 from relsad.network.systems import (
-    Distribution, 
-    PowerSystem, 
+    Distribution,
+    PowerSystem,
     Transmission,
 )
 from relsad.Time import (
@@ -29,9 +28,9 @@ from relsad.visualization.plotting import plot_topology
 def initialize_network():
 
     C1 = ManualMainController(
-            name="C1",
-            sectioning_time=Time(1, TimeUnit.HOUR),
-        )
+        name="C1",
+        sectioning_time=Time(1, TimeUnit.HOUR),
+    )
     ps = PowerSystem(C1)
 
     line_stat_dist = StatDist(
@@ -42,210 +41,208 @@ def initialize_network():
             min_val=0.5,
             max_val=2,
         ),
-        draw_flag=True,
-        get_flag=False,
     )
     fail_rate_trafo = 0.007  # 0.008
     fail_rate_line = 0.07  # 0.08
     outage_time_trafo = Time(8, TimeUnit.HOUR)
 
     B1 = Bus(
-        name="B1", 
-        coordinate=[0, 0], 
+        name="B1",
+        coordinate=[0, 0],
         fail_rate_per_year=0,
-        )
+    )
     B2 = Bus(
-        name="B2", 
-        coordinate=[0, -1], 
+        name="B2",
+        coordinate=[0, -1],
         fail_rate_per_year=fail_rate_trafo,
         outage_time=outage_time_trafo,
-        )
+    )
     B3 = Bus(
-        name="B3", 
-        coordinate=[0, -2], 
+        name="B3",
+        coordinate=[0, -2],
         fail_rate_per_year=fail_rate_trafo,
         outage_time=outage_time_trafo,
-        )
+    )
     B4 = Bus(
-        name="B4", 
-        coordinate=[0, -3], 
+        name="B4",
+        coordinate=[0, -3],
         fail_rate_per_year=fail_rate_trafo,
         outage_time=outage_time_trafo,
-        )
+    )
     B5 = Bus(
-        name="B5", 
-        coordinate=[0, -4], 
+        name="B5",
+        coordinate=[0, -4],
         fail_rate_per_year=fail_rate_trafo,
         outage_time=outage_time_trafo,
-        )
+    )
     B6 = Bus(
-        name="B6", 
-        coordinate=[0, -5], 
+        name="B6",
+        coordinate=[0, -5],
         fail_rate_per_year=fail_rate_trafo,
         outage_time=outage_time_trafo,
-        )
+    )
     B7 = Bus(
-        name="B7", 
-        coordinate=[0, -6], 
+        name="B7",
+        coordinate=[0, -6],
         fail_rate_per_year=fail_rate_trafo,
         outage_time=outage_time_trafo,
-        )
+    )
     B8 = Bus(
-        name="B8", 
-        coordinate=[0, -7], 
+        name="B8",
+        coordinate=[0, -7],
         fail_rate_per_year=fail_rate_trafo,
         outage_time=outage_time_trafo,
-        )
+    )
     B9 = Bus(
-        name="B9", 
-        coordinate=[0, -8], 
+        name="B9",
+        coordinate=[0, -8],
         fail_rate_per_year=fail_rate_trafo,
         outage_time=outage_time_trafo,
-        )
+    )
     B10 = Bus(
-        name="B10", 
-        coordinate=[0, -9], 
+        name="B10",
+        coordinate=[0, -9],
         fail_rate_per_year=fail_rate_trafo,
         outage_time=outage_time_trafo,
-        )
+    )
     B11 = Bus(
[…]
