+++ relsad/test_networks/test_10bus.py
@@ -60,8 +60,6 @@ def initialize_network(
             min_val=0.5,
             max_val=2,
         ),
-        draw_flag=True,
-        get_flag=False,
     )
 
     def num_ev_table_func(
