+++ tests/test_microgrid.py
@@ -1,5 +1,16 @@
-from relsad.network.components import Bus, Line, Disconnector, CircuitBreaker, ManualMainController
-from relsad.network.systems import PowerSystem, Distribution, Transmission, Microgrid
+from relsad.network.components import (
+    Bus,
+    Line,
+    Disconnector,
+    CircuitBreaker,
+    ManualMainController,
+)
+from relsad.network.systems import (
+    PowerSystem,
+    Distribution,
+    Transmission,
+    Microgrid,
+)
 from relsad.utils import eq
 import numpy as np
 from relsad.Time import (
@@ -14,12 +25,12 @@ def initialize_network():
 
     ps = PowerSystem(C1)
 
-    B1 = Bus(name="B1", n_customers=0, coordinate=[0,0])
-    B2 = Bus(name="B2", n_customers=1, coordinate=[0,-1])
-    B3 = Bus(name="B3", n_customers=1, coordinate=[0,-2])
-    B4 = Bus(name="B4", n_customers=1, coordinate=[-1,-3])
-    B5 = Bus(name="B5", n_customers=1, coordinate=[-1,-4])
-    B6 = Bus(name="B6", n_customers=1, coordinate=[1,-3])
+    B1 = Bus(name="B1", n_customers=0, coordinate=[0, 0])
+    B2 = Bus(name="B2", n_customers=1, coordinate=[0, -1])
+    B3 = Bus(name="B3", n_customers=1, coordinate=[0, -2])
+    B4 = Bus(name="B4", n_customers=1, coordinate=[-1, -3])
+    B5 = Bus(name="B5", n_customers=1, coordinate=[-1, -4])
+    B6 = Bus(name="B6", n_customers=1, coordinate=[1, -3])
 
     length = 1
     r = 0.5
@@ -41,75 +52,72 @@ def initialize_network():
         area=area,
         s_ref=s_ref,
         v_ref=v_ref,
-        capacity=capacity, 
+        capacity=capacity,
     )
     L2 = Line(
         name="L2",
         fbus=B2,
         tbus=B3,
-        r=r, 
+        r=r,
         x=x,
         rho=rho,
         area=area,
         s_ref=s_ref,
         v_ref=v_ref,
-        capacity=capacity, 
+        capacity=capacity,
     )
     L3 = Line(
         name="L3",
         fbus=B3,
         tbus=B4,
-        r=r, 
+        r=r,
         x=x,
         rho=rho,
         area=area,
         s_ref=s_ref,
         v_ref=v_ref,
-        capacity=capacity, 
+        capacity=capacity,
     )
     L4 = Line(
         name="L4",
         fbus=B4,
         tbus=B5,
-        r=r, 
+        r=r,
         x=x,
         rho=rho,
         area=area,
         s_ref=s_ref,
         v_ref=v_ref,
-        capacity=capacity, 
+        capacity=capacity,
     )
     L5 = Line(
         name="L5",
         fbus=B3,
         tbus=B6,
-        r=r, 
+        r=r,
         x=x,
         rho=rho,
         area=area,
         s_ref=s_ref,
         v_ref=v_ref,
-        capacity=capacity, 
+        capacity=capacity,
     )
 
-    E1 = CircuitBreaker("E1", L1)
-    EM = CircuitBreaker("EM", L3)
+    CircuitBreaker("E1", L1)
+    CircuitBreaker("EM", L3)
     tn = Transmission(ps, trafo_bus=B1)
     dn = Distribution(parent_network=tn, connected_line=L1)
-    dn.add_buses(
-        [B2, B3, B4, B5, B6]
-    )
-    dn.add_lines(
-        [L2, L3, L4, L5]
-    )
[…]
