### 2022-05-14 0c4676b
added statistical distribution for outage time
