### 2022-05-14 6e72322
removed bus outage time
