+++ relsad/test_networks/IEEE33.py
@@ -142,224 +142,192 @@ def initialize_network(
         n_customers=1,
         coordinate=[1, 0],
         fail_rate_per_year=fail_rate_trafo,
-        outage_time=outage_time_trafo,
     )
     B3 = Bus(
         "B3",
         n_customers=39,
         coordinate=[2, 0],
         fail_rate_per_year=fail_rate_trafo,
-        outage_time=outage_time_trafo,
     )
     B4 = Bus(
         "B4",
         n_customers=1,
         coordinate=[3, 0],
         fail_rate_per_year=fail_rate_trafo,
-        outage_time=outage_time_trafo,
     )
     B5 = Bus(
         "B5",
         n_customers=26,
         coordinate=[4, 0],
         fail_rate_per_year=fail_rate_trafo,
-        outage_time=outage_time_trafo,
     )
     B6 = Bus(
         "B6",
         n_customers=26,
         coordinate=[5, 0],
         fail_rate_per_year=fail_rate_trafo,
-        outage_time=outage_time_trafo,
     )
     B7 = Bus(
         "B7",
         n_customers=1,
         coordinate=[6, 0],
         fail_rate_per_year=fail_rate_trafo,
-        outage_time=outage_time_trafo,
     )
     B8 = Bus(
         "B8",
         n_customers=1,
         coordinate=[7, 0],
         fail_rate_per_year=fail_rate_trafo,
-        outage_time=outage_time_trafo,
     )
     B9 = Bus(
         "B9",
         n_customers=26,
         coordinate=[8, 0],
         fail_rate_per_year=fail_rate_trafo,
-        outage_time=outage_time_trafo,
     )
     B10 = Bus(
         "B10",
         n_customers=26,
         coordinate=[9, 0],
         fail_rate_per_year=fail_rate_trafo,
-        outage_time=outage_time_trafo,
     )
     B11 = Bus(
         "B11",
         n_customers=19,
         coordinate=[10, 0],
         fail_rate_per_year=fail_rate_trafo,
-        outage_time=outage_time_trafo,
     )
     B12 = Bus(
         "B12",
         n_customers=26,
         coordinate=[11, 0],
         fail_rate_per_year=fail_rate_trafo,
-        outage_time=outage_time_trafo,
     )
     B13 = Bus(
         "B13",
         n_customers=26,
         coordinate=[12, 0],
         fail_rate_per_year=fail_rate_trafo,
-        outage_time=outage_time_trafo,
     )
     B14 = Bus(
         "B14",
         n_customers=1,
         coordinate=[13, 0],
         fail_rate_per_year=fail_rate_trafo,
-        outage_time=outage_time_trafo,
     )
     B15 = Bus(
         "B15",
         n_customers=26,
         coordinate=[14, 0],
         fail_rate_per_year=fail_rate_trafo,
-        outage_time=outage_time_trafo,
     )
     B16 = Bus(
         "B16",
         n_customers=26,
         coordinate=[15, 0],
         fail_rate_per_year=fail_rate_trafo,
-        outage_time=outage_time_trafo,
     )
     B17 = Bus(
         "B17",
         n_customers=26,
         coordinate=[16, 0],
         fail_rate_per_year=fail_rate_trafo,
-        outage_time=outage_time_trafo,
     )
     B18 = Bus(
         "B18",
         n_customers=39,
         coordinate=[17, 0],
         fail_rate_per_year=fail_rate_trafo,
-        outage_time=outage_time_trafo,
     )
     B19 = Bus(
         "B19",
         n_customers=39,
         coordinate=[2, 1],
         fail_rate_per_year=fail_rate_trafo,
-        outage_time=outage_time_trafo,
     )
     B20 = Bus(
         "B20",
[…]
