+++ docs/source/API/utils.rst
@@ -1,0 +1,9 @@
+=========
+Utilities
+=========
+
+.. autosummary::
+    :toctree: _autosummary
+    :template: module.rst
+
+    relsad.utils
