+++ docs/source/API/dist_and_tables.rst
@@ -1,0 +1,28 @@
+========================
+Distributions and tables
+========================
+
+''''''''
+StatDist
+''''''''
+
+.. autosummary::
+    :toctree: _autosummary
+    :template: class.rst
+
+    relsad.StatDist.StatDist
+    relsad.StatDist.StatDistType
+    relsad.StatDist.UniformParameters
+    relsad.StatDist.NormalParameters
+    relsad.StatDist.GammaParameters
+
+
+'''''
+Table
+'''''
+
+.. autosummary::
+    :toctree: _autosummary
+    :template: class.rst
+
+    relsad.Table.Table
