+++ docs/source/API/time.rst
@@ -1,0 +1,24 @@
+==================
+Time and timestamp
+==================
+
+''''
+Time
+''''
+
+.. autosummary::
+    :toctree: _autosummary
+    :template: class.rst
+
+    relsad.Time.Time
+
+
+'''''''''
+TimeStamp
+'''''''''
+
+.. autosummary::
+    :toctree: _autosummary
+    :template: class.rst
+
+    relsad.Time.TimeStamp
