### 2022-05-16 4bbbfbe
added creation of random instance if none provided
