+++ pyproject.toml
@@ -20,6 +20,7 @@ sphinx-book-theme = "^0.3.2"
 Sphinx = "^3.5.2"
 sphinxcontrib-napoleon = "^0.7"
 pytest = "^7.1.1"
+pytest-cov = "^3.0.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
