### 2022-05-18 ed64c46
Bug fix: Added check for number in Time initialization
