### 2022-05-28 6d2fa7d
distinguished outage time and repair time
