+++ relsad/network/components/Bus.py
@@ -92,8 +92,8 @@ class Bus(Component):
         The failure rate per year for the transformer at the bus
     trafo_failed : bool
         Failure status of the transformer
-    outage_time_dist : StatDist
-        The outage time of the transformer at the bus [hours/fault]
+    repair_time_dist : StatDist
+        The repair time of the transformer at the bus [hours/fault]
     remaining_outage_time : Time
         The remaining outage time of the line
     acc_outage_time : Time
@@ -140,8 +140,8 @@ class Bus(Component):
         Sets the bus load and cost in MW based on load and cost profiles in the current increment
     get_load()
         Retuns the current load at the bus in MW
-    draw_outage_time(dt)
-        Decides and returns the outage time of the trafo based on a statistical distribution
+    draw_repair_time(dt)
+        Decides and returns the repair time of the trafo based on a statistical distribution
     trafo_fail(dt)
         Sets the transformer status to failed, load and generation at the node are set to zero
     trafo_not_fail()
@@ -208,7 +208,7 @@ class Bus(Component):
         s_ref: float = 1,  # MVA
         is_slack: bool = False,
         fail_rate_per_year: float = 0.0,
-        outage_time_dist: StatDist = StatDist(
+        repair_time_dist: StatDist = StatDist(
             stat_dist_type=StatDistType.UNIFORM_FLOAT,
             parameters=UniformParameters(
                 min_val=0.0,
@@ -254,7 +254,7 @@ class Bus(Component):
 
         ## Reliabilility attributes
         self.fail_rate_per_year = fail_rate_per_year  # failures per year
-        self.outage_time_dist = outage_time_dist
+        self.repair_time_dist = repair_time_dist
         self.acc_outage_time = Time(0)
         self.avg_fail_rate = 0
         self.avg_outage_time = Time(0)
@@ -484,9 +484,9 @@ class Bus(Component):
         """
         return self.pload, self.qload
 
-    def draw_outage_time(self, dt: Time):
+    def draw_repair_time(self, dt: Time):
         """
-        Decides and returns the outage time of the trafo based on a statistical distribution
+        Decides and returns the repair time of the trafo based on a statistical distribution
 
         Parameters
         ----------
@@ -499,7 +499,7 @@ class Bus(Component):
 
         """
         return Time(
-            self.outage_time_dist.draw(
+            self.repair_time_dist.draw(
                 random_instance=self.ps_random,
                 size=1,
             )[0],
@@ -521,7 +521,7 @@ class Bus(Component):
 
         """
         self.trafo_failed = True
-        self.remaining_outage_time = self.draw_outage_time(dt)
+        self.remaining_outage_time = self.draw_repair_time(dt)
         self.shed_load(dt)
         if self.prod is not None:
             self.prod.reset_prod()
