### 2022-05-28 237bc83
improved random instance interface
