### 2022-05-28 da0a313
added handling of corner case cb line failure combination
