+++ tests/test_network_behavior_limited_support_after_sectioning_time.py
@@ -8,6 +8,12 @@ from relsad.Time import (
     TimeUnit,
 )
 
+from relsad.StatDist import (
+    StatDist,
+    StatDistType,
+    UniformParameters,
+)
+
 
 class MyError(Exception):
     def __init__(self, m):
@@ -1134,6 +1140,230 @@ def test_L3_and_L5_fail():
     assert ps.get_comp("L5b").is_open is True
     assert ps.get_comp("L6a").is_open is False
     assert ps.get_comp("L6b").is_open is False
+    assert ps.get_comp("E2").is_open is False
+    assert ps.get_comp("L7a").is_open is False
+    assert ps.get_comp("L7b").is_open is False
+    assert ps.get_comp("L7c").is_open is False
+    assert ps.get_comp("ML1a").is_open is False
+    assert ps.get_comp("ML1b").is_open is False
+    assert ps.get_comp("ML2a").is_open is False
+    assert ps.get_comp("ML2b").is_open is False
+
+
+def test_L2_L1_fail():
+    ps = initialize_network(
+        include_ICT=False,
+        microgrid_mode=MicrogridMode.LIMITED_SUPPORT,
+        fail_rate_line=0,
+        line_stat_dist=StatDist(
+            stat_dist_type=StatDistType.UNIFORM_FLOAT,
+            parameters=UniformParameters(
+                min_val=2,
+                max_val=2,
+            ),
+        ),
+    )
+
+    # Create sections
+    ps.create_sections()
+
+    # Simulate failure at time 0
+    # Time = 0
+    curr_time = Time(0)
+    dt = Time(1, TimeUnit.HOUR)
+
+    ps.update_fail_status(dt)
+
+    ps.get_comp("L2").fail(dt)
+
+    ps.controller.run_control_loop(curr_time, dt)
+
+    find_sub_systems(ps, curr_time)
+
+    assert ps.get_comp("L1").connected is False
+    assert ps.get_comp("L2").connected is True
+    assert ps.get_comp("L3").connected is True
+    assert ps.get_comp("L4").connected is True
+    assert ps.get_comp("L5").connected is True
+    assert ps.get_comp("L6").connected is False
+    assert ps.get_comp("L7").connected is False
+    assert ps.get_comp("ML1").connected is True
+    assert ps.get_comp("ML2").connected is True
+    assert ps.get_comp("E1").is_open is True
+    assert ps.get_comp("L1a").is_open is True
+    assert ps.get_comp("L1b").is_open is True
+    assert ps.get_comp("L1c").is_open is True
+    assert ps.get_comp("L2a").is_open is False
+    assert ps.get_comp("L2b").is_open is False
+    assert ps.get_comp("L3a").is_open is False
+    assert ps.get_comp("L3b").is_open is False
+    assert ps.get_comp("L4a").is_open is False
+    assert ps.get_comp("L4b").is_open is False
+    assert ps.get_comp("L5a").is_open is False
+    assert ps.get_comp("L5b").is_open is False
+    assert ps.get_comp("L6a").is_open is True
+    assert ps.get_comp("L6b").is_open is True
+    assert ps.get_comp("E2").is_open is True
+    assert ps.get_comp("L7a").is_open is True
+    assert ps.get_comp("L7b").is_open is True
+    assert ps.get_comp("L7c").is_open is True
+    assert ps.get_comp("ML1a").is_open is False
+    assert ps.get_comp("ML1b").is_open is False
+    assert ps.get_comp("ML2a").is_open is False
+    assert ps.get_comp("ML2b").is_open is False
+
+    # Time = 1
+    ps.update_fail_status(dt)
+    ps.get_comp("L1").fail(dt)
+
+    curr_time = Time(1)
+    ps.controller.run_control_loop(curr_time, dt)
+
[…]
