### 2022-05-28 7de7e76
fixed bug regarding section removal
