+++ docs/source/installation.rst
@@ -4,23 +4,23 @@
 Installation
 ============
 
+..
+    Stable release
+    --------------
 
-Stable release
---------------
+    To install relsad, run this command in your terminal:
 
-To install relsad, run this command in your terminal:
+    .. code-block:: console
 
-.. code-block:: console
+        $ pip install -u relsad
 
-    $ pip install -u relsad
+    This is the preferred method to install relsad, as it will always install the most recent stable release.
 
-This is the preferred method to install relsad, as it will always install the most recent stable release.
+    If you don't have `pip`_ installed, this `Python installation guide`_ can guide
+    you through the process.
 
-If you don't have `pip`_ installed, this `Python installation guide`_ can guide
-you through the process.
-
-.. _pip: https://pip.pypa.io
-.. _Python installation guide: http://docs.python-guide.org/en/latest/starting/installation/
+    .. _pip: https://pip.pypa.io
+    .. _Python installation guide: http://docs.python-guide.org/en/latest/starting/installation/
 
 
 From sources
