+++ relsad/network/components/Disconnector.py
@@ -96,6 +96,25 @@ class Disconnector(Component):
         fail_rate: float = 0.014,
         outage_time: Time = Time(1, TimeUnit.HOUR),
     ):
+
+        # Verify input
+        if line is None:
+            raise Exception(
+                "Disconnector must be connected to a Line"
+            )
+        if line.parent_network is not None:
+            raise Exception(
+                "Disconnector must be created before the line is connected to a network"
+            )
+        if bus is None:
+            raise Exception(
+                "Disconnector must be connected to a Bus"
+            )
+        if fail_rate < 0:
+            raise Exception(
+                "The failure rate must be positive"
+            )
+
         self.name = name
         self.initial_state = is_open
         self.is_open = is_open
