+++ relsad/visualization/plotting.py
@@ -8,6 +8,7 @@ from relsad.network.components import (
     CircuitBreaker,
     IntelligentSwitch,
     Sensor,
+    EVPark,
 )
 import os
 
@@ -21,6 +22,7 @@ def plot_topology(
     disconnector_text: bool = False,
     intelligent_switch_text: bool = False,
     sensor_text: bool = False,
+    text_dx: tuple = (0, -0.1),
     **kwargs
 ):
 
@@ -55,14 +57,24 @@ def plot_topology(
     None
 
     """
+    # Handle keyword arguments
+
+    left = kwargs["left"] if "left" in kwargs else 0.02
+    right = kwargs["right"] if "right" in kwargs else 0.98
+    bottom = kwargs["bottom"] if "bottom" in kwargs else 0.1
+    top = kwargs["top"] if "top" in kwargs else 0.75
+
+    text_size = kwargs["text_size"] if "text_size" in kwargs else 6
+
+    # Clear used keyword arguments
+    kwargs.pop("left", None)
+    kwargs.pop("right", None)
+    kwargs.pop("bottom", None)
+    kwargs.pop("top", None)
+    kwargs.pop("text_size", None)
 
     fig, ax = plt.subplots(**kwargs)
-    left = 0.02
-    right = 0.98
-    bottom = 0.1
-    top = 0.75
-
-    text_size = 6
+    
 
     fig.subplots_adjust(
         left=left,
@@ -74,44 +86,71 @@ def plot_topology(
     )
     legends = {}
     for bus in buses:
-        _plot_bus(ax, bus, text=bus_text, text_size=text_size)
+        _plot_bus(
+            ax=ax,
+            bus=bus,
+            text=bus_text,
+            text_size=text_size,
+            text_dx=text_dx,
+        )
         legends["Bus"] = Bus.handle
+        if bus.ev_park is not None:
+            legends["EV park"] = EVPark.handle
     for line in lines:
-        _plot_line(ax, line, text=line_text, text_size=text_size)
+        _plot_line(
+            ax=ax,
+            line=line,
+            text=line_text,
+            text_size=text_size,
+        )
         legends["Line"] = Line.handle
         if line.circuitbreaker is not None:
             _plot_circuitbreaker(
-                ax, line, text=circuitbreaker_text, text_size=text_size
+                ax=ax,
+                line=line,
+                text=circuitbreaker_text,
+                text_size=text_size,
             )
             legends["Circuit breaker"] = CircuitBreaker.handle
             for discon in line.circuitbreaker.disconnectors:
                 _plot_disconnector(
-                    ax, discon, text=disconnector_text, text_size=text_size
+                    ax=ax,
+                    discon=discon,
+                    text=disconnector_text,
+                    text_size=text_size,
                 )
                 legends["Disconnector"] = Disconnector.handle
                 if discon.intelligent_switch:
                     _plot_intelligent_switch(
-                        ax,
-                        discon,
+                        ax=ax,
+                        discon=discon,
                         text=intelligent_switch_text,
                         text_size=text_size,
                     )
                     legends["Intelligent switch"] = IntelligentSwitch.handle
         for discon in line.disconnectors:
             _plot_disconnector(
-                ax, discon, text=disconnector_text, text_size=text_size
+                ax=ax,
+                discon=discon,
+                text=disconnector_text,
+                text_size=text_size,
             )
             legends["Disconnector"] = Disconnector.handle
[…]
