### 2022-05-30 aa5c7b4
fixed bug with test network
