### 2022-05-30 a075492
updated README file type in setup.py
