+++ setup.py
@@ -1,14 +1,14 @@
 import setuptools
 
-with open("README.md", "r") as fh:
+with open("README.rst", "r") as fh:
     long_description = fh.read()
 
 setuptools.setup(
-    name="relsad",  # Replace with your own username
+    name="relsad",
     version="0.0.1",
     author="Stine Fleischer Myhre",
     author_email="stine.f.myhre@ntnu.no",
-    description="A small example package",
+    description="A Python package for reliability assessment of modern distribution systems.",
     long_description=long_description,
     long_description_content_type="text/markdown",
     url="https://github.com/stinefm/STNetwork",
