### 2022-05-31 9982e9c
Reformulated index descriptions
