### 2022-06-07 #4391 @fraukewiese
@nesnoj : Would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
The submission under review here is [RELSAD: A Python package for reliability assessment of modern distribution systems](https://github.com/stinefm/relsad)
