+++ relsad/network/components/Disconnector.py
@@ -99,21 +99,15 @@ class Disconnector(Component):
 
         # Verify input
         if line is None:
-            raise Exception(
-                "Disconnector must be connected to a Line"
-            )
+            raise Exception("Disconnector must be connected to a Line")
         if line.parent_network is not None:
             raise Exception(
                 "Disconnector must be created before the line is connected to a network"
             )
         if bus is None:
-            raise Exception(
-                "Disconnector must be connected to a Bus"
-            )
+            raise Exception("Disconnector must be connected to a Bus")
         if fail_rate < 0:
-            raise Exception(
-                "The failure rate must be positive"
-            )
+            raise Exception("The failure rate must be positive")
 
         self.name = name
         self.initial_state = is_open
