### 2022-06-14 b1c1f16
added check for num_cars key
