+++ relsad/reliability/indices.py
@@ -152,6 +152,8 @@ def EV_Interruption(network: Network):
     average_num_cars = sum(
         [
             np.mean(list(ev_park.history["num_cars"].values()))
+            if "num_cars" in ev_park.history
+            else 0
             for ev_park in network.ev_parks
         ]
     )
