### 2022-06-18 82df5ee
restructured examples and test networks
