### 2022-06-18 0239dd0
added check for prod data
