### 2022-06-19 190b546
added BatteryType and BatteryState
