### 2022-06-19 dfc0640
added verify_component_setup
