### 2022-06-19 183cb8e
added docstrings to enums
