+++ relsad/network/components/Controller.py
@@ -4,6 +4,21 @@ from relsad.Time import Time
 
 
 class ControllerState(Enum):
+    """
+    Controller state
+
+    Attributes
+    ----------
+    OK : int
+        The controller is up and running
+    SOFTWARE_FAIL : int
+        A software fail has occurred
+    HARDWARE_FAIL : int
+        A hardware fail has occurred
+    REPAIR : int
+        The controller is being repaired
+    """
+
     OK = 1
     SOFTWARE_FAIL = 2
     HARDWARE_FAIL = 3
