+++ relsad/network/components/EVPark.py
@@ -60,10 +60,6 @@ class EVPark(Component):
         Fraction of interruption experienced by the EV park
     acc_num_cars : int
         Accumulated number of cars in EV park
-    num_car_draws : int
-        Number of car draws in the EV park
-    acc_num_interruptions : float
-        Accumulated number of interruptions experienced by the EV park
     acc_exp_interruptions : float
         Accumulated experienced interruptions in the EV park
     curr_exp_interruptions : float
@@ -189,8 +185,6 @@ class EVPark(Component):
         self.num_consecutive_interruptions = 0
         self.park_interruption_fraction = 0
         self.acc_num_cars = 0
-        self.num_car_draws = 0
-        self.acc_num_interruptions = 0
         self.curr_exp_interruptions = 0
         self.acc_exp_interruptions = 0
         self.curr_exp_car_interruptions = 0
@@ -292,7 +286,6 @@ class EVPark(Component):
         """
         if fail_duration == dt:
             self.draw_current_state(hour_of_day)
-            self.num_car_draws += 1
             self.acc_num_cars += self.num_cars
 
         # Update car batteries based on system balance
@@ -406,7 +399,6 @@ class EVPark(Component):
         self.history["charge"] = {}
         self.history["num_cars"] = {}
         self.history["park_interruption_fraction"] = {}
-        self.history["acc_num_interruptions"] = {}
         self.history["acc_exp_interruptions"] = {}
         self.history["acc_exp_car_interruptions"] = {}
         self.history["acc_interruption_duration"] = {}
@@ -447,7 +439,6 @@ class EVPark(Component):
             self.curr_interruption_duration += dt
         else:
             if self.num_consecutive_interruptions >= 1:
-                self.acc_num_interruptions += 1
                 self.acc_exp_interruptions += (
                     self.curr_exp_interruptions
                     / self.num_consecutive_interruptions
@@ -472,9 +463,6 @@ class EVPark(Component):
             self.history["park_interruption_fraction"][
                 curr_time
             ] = self.park_interruption_fraction
-            self.history["acc_num_interruptions"][
-                curr_time
-            ] = self.acc_num_interruptions
             self.history["acc_exp_interruptions"][
                 curr_time
             ] = self.acc_exp_interruptions
@@ -582,8 +570,6 @@ class EVPark(Component):
         self.num_consecutive_interruptions = 0
         self.park_interruption_fraction = 0
         self.acc_num_cars = 0
-        self.num_car_draws = 0
-        self.acc_num_interruptions = 0
         self.curr_exp_interruptions = 0
         self.acc_exp_interruptions = 0
         self.curr_exp_car_interruptions = 0
