### 2022-06-20 f0815e2
undo removal of acc_num_interruptions
