### 2022-06-27 #4516 @editorialbot
[REVIEW]: RELSAD: A Python package for reliability assessment of modern distribution systems
**Submitting author:** @stinefm (<a href="http://orcid.org/0000-0002-2283-1724">Stine Fleischer Myhre</a>)
**Repository:** https://github.com/stinefm/relsad
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.14
