+++ joss/paper.bib
@@ -1,5 +1,5 @@
 @book{Billinton:1992,
-  	url = {https://doi.org/10.1007/978-1-4615-7728-7},
+  	url = {10.1007/978-1-4615-7728-7},
   	Author = {{Billinton}, R. and {Allan}, R. N.},
   	Booktitle = {Reliability evaluation of engineering systems: First Edition, by Roy Billinton and Ronald N. Allan.~ISBN 978-1-4615-7728-7 (eBook).~Published by Springer, Springer, Boston, MA, 1992.},
   	Publisher = {Springer},
@@ -11,7 +11,7 @@
 @article{Kjølle:1992,
   title={RELRAD-an analytical approach for distribution system reliability assessment},
   author={{Kjølle}, G. and {Sand}, K.},
-  url = {http://doi.org/10.1109/61.127084},
+  doi = {10.1109/tdc.1991.169586},
   journal={IEEE Transactions on Power Delivery},
   volume={7},
   number={2},
@@ -23,7 +23,7 @@
 @article{escalera:2018,
   title={A survey of reliability assessment techniques for modern distribution networks},
   author={Escalera, Alberto and Hayes, Barry and Prodanovi{\'c}, Milan},
-  url={https://doi.org/10.1016/j.rser.2018.02.031},
+  doi={10.1016/j.rser.2018.02.031},
   journal={Renewable and Sustainable Energy Reviews},
   volume={91},
   pages={344--357},
@@ -36,7 +36,7 @@
 @article{celli:2013,
   title={Reliability assessment in smart distribution networks},
   author={Celli, Gianni and Ghiani, Emilio and Pilo, Fabrizio and Soma, Gian Giuseppe},
-  url={https://doi.org/10.1016/j.epsr.2013.07.001},
+  doi={10.1016/j.epsr.2013.07.001},
   journal={Electric Power Systems Research},
   volume={104},
   pages={164--175},
@@ -47,7 +47,7 @@
 @article{de:2017,
   title={Reliability assessment of microgrids with local and mobile generation, time-dependent profiles, and intraday reconfiguration},
   author={de Quevedo, Pilar Meneses and Contreras, Javier and Mazza, Andrea and Chicco, Gianfranco and Porumb, Radu},
-  url={https://doi.org/10.1109/TIA.2017.2752685},
+  doi={10.1109/TIA.2017.2752685},
   journal={IEEE Transactions on Industry Applications},
   volume={54},
   number={1},
@@ -59,84 +59,11 @@
 @article{borges:2012,
   title={An overview of reliability models and methods for distribution systems with renewable energy distributed generation},
   author={Borges, Carmen Lucia Tancredo},
-  url={https://doi.org/10.1016/j.rser.2012.03.055},
+  doi={10.1016/j.rser.2012.03.055},
   journal={Renewable and sustainable energy reviews},
   volume={16},
   number={6},
   pages={4008--4015},
   year={2012},
   publisher={Elsevier}
-}
-
-@inproceedings{chaudonneret_representation_2012,
-	title = {Representation of the influence of telecommunications on electrical distribution network reliability},
-	url = {http://doi.org/10.1109/SmartGridComm.2012.6485993},
-	booktitle = {2012 {IEEE} {Third} {International} {Conference} on {Smart} {Grid} {Communications} ({SmartGridComm})},
-	author = {Chaudonneret, T. and Decroix, H. and McDonald, J. D. F.},
-	month = nov,
-	year = {2012},
[…]
