+++ relsad/simulation/Simulation.py
@@ -152,12 +152,12 @@ class Simulation:
             update_sub_system_slack(self.power_system)
             ## Load flow
             for sub_system in self.power_system.sub_systems:
+                ## Update batteries and history
+                sub_system.update_batteries(self.fail_duration, dt)
                 ## Update EV parks
                 sub_system.update_ev_parks(
                     self.fail_duration, dt, start_time, curr_time
                 )
-                ## Update batteries and history
-                sub_system.update_batteries(self.fail_duration, dt)
                 ## Run load flow
                 sub_system.reset_load_flow_data()
                 if sub_system.slack is not None:
