+++ docs/source/theory/Components/controller.rst
@@ -2,6 +2,6 @@
 Controller
 ======================
 
-In RELSAD, the power system should be initialized with a main controller. This could either be a manuall controller which repesent a more traditional method of controlling the system where multiple actions are performed manually or it could be a smarter controller with a higher share of automation. 
-For each child network in the power system, the main controller will create and allocated a controller for of the  networks. 
+In RELSAD, the power system should be initialized with a main controller. This could either be a manual controller which represent a more traditional method of controlling the system where multiple actions are performed manually, or it could be a smarter controller with a higher share of automation. 
+For each child network in the power system, the main controller will create and allocated a controller for of the networks. 
 The control system is made to simulate a central control strategy similar to a control center at a system operator. 
