+++ relsad/reliability/indices.py
@@ -146,21 +146,18 @@ def EV_Interruption(network: Network):
         The average number of interruptions per EV car for grid support
 
     """
-    interrupted_cars = sum(
+    interruptions_x_num_cars = sum(
         [
-            ev_park.acc_exp_car_interruptions
+            ev_park.acc_exp_interruptions * ev_park.acc_num_cars
             for ev_park in network.ev_parks
         ]
     )
     total_num_cars = sum(
-        [
-            ev_park.acc_num_cars
-            for ev_park in network.ev_parks
-        ]
+        [ev_park.acc_num_cars for ev_park in network.ev_parks]
     )
     if total_num_cars == 0:
         return 0
-    ev_interruption = interrupted_cars / total_num_cars
+    ev_interruption = interruptions_x_num_cars / total_num_cars
     return ev_interruption
 
 
