### 2022-07-18 b2dab32
added ASAI and ASUI indices
