+++ relsad/simulation/monte_carlo/history.py
@@ -1,6 +1,6 @@
 import os
 import copy
-from relsad.Time import TimeUnit
+from relsad.Time import Time
 from relsad.network.systems import (
     PowerSystem,
 )
@@ -28,7 +28,9 @@ from relsad.reliability.indices import (
     SAIFI,
     SAIDI,
     CAIDI,
-    EENS,
+    ASAI,
+    ASUI,
+    ENS,
     EV_Index,
     EV_Interruption,
     EV_Duration,
@@ -57,7 +59,9 @@ def plot_network_monte_carlo_history(power_system: PowerSystem, save_dir: str):
         "SAIFI",
         "SAIDI",
         "CAIDI",
-        "EENS",
+        "ASAI",
+        "ASUI",
+        "ENS",
         "EV_Index",
         "EV_Interruption",
         "EV_Duration",
@@ -113,7 +117,9 @@ def save_network_monte_carlo_history(
         "SAIFI",
         "SAIDI",
         "CAIDI",
-        "EENS",
+        "ASAI",
+        "ASUI",
+        "ENS",
         "EV_Index",
         "EV_Interruption",
         "EV_Duration",
@@ -196,7 +202,9 @@ def initialize_monte_carlo_history(power_system: PowerSystem):
         "SAIFI",
         "SAIDI",
         "CAIDI",
-        "EENS",
+        "ASAI",
+        "ASUI",
+        "ENS",
         "EV_Index",
         "EV_Interruption",
         "EV_Duration",
@@ -308,7 +316,10 @@ def save_iteration_history(power_system: PowerSystem, it: int, save_dir: str):
 
 
 def update_monte_carlo_power_system_history(
-    power_system: PowerSystem, it: int, time_unit: TimeUnit, save_dict: dict
+    power_system: PowerSystem,
+    it: int,
+    current_time: Time,
+    save_dict: dict,
 ):
     """
     Updates the history dictionary from the Monte Carlo simulation
@@ -319,8 +330,8 @@ def update_monte_carlo_power_system_history(
         A power system element
     it : int
         The iteration number
-    time_unit : TimeUnit
-        A time unit (hour, seconds, ect.)
+    current_time : Time
+        Current time
     save_dict : dict
         Dictionary with simulation results
 
@@ -334,24 +345,29 @@ def update_monte_carlo_power_system_history(
         "acc_p_load_shed": power_system.acc_p_load_shed,
         "acc_q_load_shed": power_system.acc_q_load_shed,
         "SAIFI": SAIFI(power_system),
-        "SAIDI": SAIDI(power_system, time_unit),
-        "CAIDI": CAIDI(power_system, time_unit),
-        "EENS": EENS(power_system),
+        "SAIDI": SAIDI(power_system),
+        "CAIDI": CAIDI(power_system),
+        "ASAI": ASAI(power_system, current_time),
+        "ASUI": ASUI(power_system, current_time),
+        "ENS": ENS(power_system),
         "EV_Index": EV_Index(power_system),
         "EV_Interruption": EV_Interruption(power_system),
-        "EV_Duration": EV_Duration(power_system, time_unit),
+        "EV_Duration": EV_Duration(power_system),
     }
     for state_var, value in network_state_dict.items():
         save_dict[power_system.name][state_var][it] = value
     save_dict = update_monte_carlo_child_network_history(
-        power_system, it, time_unit, save_dict
+        power_system, it, current_time, save_dict
     )
     save_dict = update_monte_carlo_comp_history(power_system, it, save_dict)
     return save_dict
 
 
 def update_monte_carlo_child_network_history(
-    power_system: PowerSystem, it: int, time_unit: TimeUnit, save_dict: dict
+    power_system: PowerSystem,
[…]
