### 2022-07-18 ad91631
changed EV_int index formulation
