### 2022-07-29 45c7aad
added ICT network, node and line
