+++ relsad/simulation/sequence/history.py
@@ -42,20 +42,20 @@ def update_history(
     """
     for network in power_system.child_network_list:
         for bus in network.buses:
-            network.p_load_shed += bus.p_load_shed_stack
-            network.q_load_shed += bus.q_load_shed_stack
-        power_system.p_load_shed += network.p_load_shed
-        power_system.q_load_shed += network.q_load_shed
-        network.acc_p_load_shed += network.p_load_shed
-        network.acc_q_load_shed += network.q_load_shed
-    power_system.acc_p_load_shed += power_system.p_load_shed
-    power_system.acc_q_load_shed += power_system.q_load_shed
+            network.p_energy_shed += bus.p_energy_shed_stack
+            network.q_energy_shed += bus.q_energy_shed_stack
+        power_system.p_energy_shed += network.p_energy_shed
+        power_system.q_energy_shed += network.q_energy_shed
+        network.acc_p_energy_shed += network.p_energy_shed
+        network.acc_q_energy_shed += network.q_energy_shed
+    power_system.acc_p_energy_shed += power_system.p_energy_shed
+    power_system.acc_q_energy_shed += power_system.q_energy_shed
     if save_flag:
         power_system_state_dict = {
-            "p_load_shed": power_system.p_load_shed,
-            "q_load_shed": power_system.p_load_shed,
-            "acc_p_load_shed": power_system.acc_p_load_shed,
-            "acc_q_load_shed": power_system.acc_q_load_shed,
+            "p_energy_shed": power_system.p_energy_shed,
+            "q_energy_shed": power_system.p_energy_shed,
+            "acc_p_energy_shed": power_system.acc_p_energy_shed,
+            "acc_q_energy_shed": power_system.acc_q_energy_shed,
             "p_load": power_system.get_system_load()[0],
             "q_load": power_system.get_system_load()[1],
         }
@@ -63,20 +63,20 @@ def update_history(
             power_system.history[state_var][curr_time] = value
         for network in power_system.child_network_list:
             network_state_dict = {
-                "p_load_shed": network.p_load_shed,
-                "q_load_shed": network.q_load_shed,
-                "acc_p_load_shed": network.acc_p_load_shed,
-                "acc_q_load_shed": network.acc_q_load_shed,
+                "p_energy_shed": network.p_energy_shed,
+                "q_energy_shed": network.q_energy_shed,
+                "acc_p_energy_shed": network.acc_p_energy_shed,
+                "acc_q_energy_shed": network.acc_q_energy_shed,
                 "p_load": network.get_system_load()[0],
                 "q_load": network.get_system_load()[1],
             }
             for state_var, value in network_state_dict.items():
                 network.history[state_var][curr_time] = value
-    power_system.p_load_shed = 0
-    power_system.q_load_shed = 0
+    power_system.p_energy_shed = 0
+    power_system.q_energy_shed = 0
     for network in power_system.child_network_list:
-        network.p_load_shed = 0
-        network.q_load_shed = 0
+        network.p_energy_shed = 0
+        network.q_energy_shed = 0
     for comp in power_system.comp_list:
         comp.update_history(prev_time, curr_time, save_flag)
     power_system.controller.update_history(prev_time, curr_time, save_flag)
@@ -351,16 +351,16 @@ def plot_bus_history(buses, save_dir: str):
         "qprod",
         "remaining_outage_time",
         "trafo_failed",
-        "p_load_shed_stack",
-        "q_load_shed_stack",
+        "p_energy_shed_stack",
+        "q_energy_shed_stack",
         "voang",
         "vomag",
     ]
[…]
