+++ docs/source/API/network/systems.rst
@@ -5,10 +5,10 @@ systems
     :toctree: _autosummary
     :template: class.rst
 
-    relsad.network.systems.Distribution.Distribution
-    relsad.network.systems.ICT.ICT
-    relsad.network.systems.Microgrid.Microgrid
-    relsad.network.systems.Network.Network
+    relsad.network.systems.PowerNetwork.PowerNetwork
     relsad.network.systems.PowerSystem.PowerSystem
-    relsad.network.systems.SubSystem.SubSystem
     relsad.network.systems.Transmission.Transmission
+    relsad.network.systems.Distribution.Distribution
+    relsad.network.systems.Microgrid.Microgrid
+    relsad.network.systems.SubSystem.SubSystem
+    relsad.network.systems.ICTNetwork.ICTNetwork
