+++ tests/test_circuitbreaker.py
@@ -11,12 +11,11 @@ def test_close():
     B2 = Bus("B2")
     L1 = Line("L1", B1, B2, 0.5, 0.5)
     CB = CircuitBreaker("CB", L1)
-    D1 = Disconnector("D1", L1, B1, CB)
 
     CB.close()
 
     assert CB.is_open is False
-    assert D1.is_open is False
+    assert L1.connected is True
 
 
 def test_open():
@@ -24,9 +23,8 @@ def test_open():
     B2 = Bus("B2")
     L1 = Line("L1", B1, B2, 0.5, 0.5)
     CB = CircuitBreaker("CB", L1)
-    D1 = Disconnector("D1", L1, B1, CB)
 
     CB.open()
 
     assert CB.is_open is True
-    assert D1.is_open is True
+    assert L1.connected is False
