+++ tests/test_network_behavior_full_support_before_sectioning_time.py
@@ -49,7 +49,6 @@ def test_B2_trafo_fail():
     assert ps.get_comp("E1").is_open is False
     assert ps.get_comp("L1a").is_open is False
     assert ps.get_comp("L1b").is_open is False
-    assert ps.get_comp("L1c").is_open is False
     assert ps.get_comp("L2a").is_open is False
     assert ps.get_comp("L2b").is_open is False
     assert ps.get_comp("L3a").is_open is False
@@ -63,7 +62,6 @@ def test_B2_trafo_fail():
     assert ps.get_comp("E2").is_open is False
     assert ps.get_comp("L7a").is_open is False
     assert ps.get_comp("L7b").is_open is False
-    assert ps.get_comp("L7c").is_open is False
     assert ps.get_comp("ML1a").is_open is False
     assert ps.get_comp("ML1b").is_open is False
     assert ps.get_comp("ML2a").is_open is False
@@ -102,7 +100,6 @@ def test_B3_trafo_fail():
     assert ps.get_comp("E1").is_open is False
     assert ps.get_comp("L1a").is_open is False
     assert ps.get_comp("L1b").is_open is False
-    assert ps.get_comp("L1c").is_open is False
     assert ps.get_comp("L2a").is_open is False
     assert ps.get_comp("L2b").is_open is False
     assert ps.get_comp("L3a").is_open is False
@@ -116,7 +113,6 @@ def test_B3_trafo_fail():
     assert ps.get_comp("E2").is_open is False
     assert ps.get_comp("L7a").is_open is False
     assert ps.get_comp("L7b").is_open is False
-    assert ps.get_comp("L7c").is_open is False
     assert ps.get_comp("ML1a").is_open is False
     assert ps.get_comp("ML1b").is_open is False
     assert ps.get_comp("ML2a").is_open is False
@@ -155,7 +151,6 @@ def test_B4_trafo_fail():
     assert ps.get_comp("E1").is_open is False
     assert ps.get_comp("L1a").is_open is False
     assert ps.get_comp("L1b").is_open is False
-    assert ps.get_comp("L1c").is_open is False
     assert ps.get_comp("L2a").is_open is False
     assert ps.get_comp("L2b").is_open is False
     assert ps.get_comp("L3a").is_open is False
@@ -169,7 +164,6 @@ def test_B4_trafo_fail():
     assert ps.get_comp("E2").is_open is False
     assert ps.get_comp("L7a").is_open is False
     assert ps.get_comp("L7b").is_open is False
-    assert ps.get_comp("L7c").is_open is False
     assert ps.get_comp("ML1a").is_open is False
     assert ps.get_comp("ML1b").is_open is False
     assert ps.get_comp("ML2a").is_open is False
@@ -208,7 +202,6 @@ def test_B5_trafo_fail():
     assert ps.get_comp("E1").is_open is False
     assert ps.get_comp("L1a").is_open is False
     assert ps.get_comp("L1b").is_open is False
-    assert ps.get_comp("L1c").is_open is False
     assert ps.get_comp("L2a").is_open is False
     assert ps.get_comp("L2b").is_open is False
     assert ps.get_comp("L3a").is_open is False
@@ -222,7 +215,6 @@ def test_B5_trafo_fail():
     assert ps.get_comp("E2").is_open is False
     assert ps.get_comp("L7a").is_open is False
     assert ps.get_comp("L7b").is_open is False
-    assert ps.get_comp("L7c").is_open is False
     assert ps.get_comp("ML1a").is_open is False
     assert ps.get_comp("ML1b").is_open is False
     assert ps.get_comp("ML2a").is_open is False
@@ -261,7 +253,6 @@ def test_B6_trafo_fail():
[…]
