+++ pytest.ini
@@ -31,4 +31,5 @@ testpaths =
     tests/test_simulation.py
     tests/test_system_config.py
     tests/test_topology_ICT_dfs.py
+    tests/test_utils.py
 
