+++ relsad/network/components/MicrogridController.py
@@ -13,6 +13,7 @@ from relsad.utils import (
     convert_yearly_fail_rate,
     unique,
 )
+from relsad.topology.ICT.dfs import is_connected
 from relsad.Time import (
     Time,
     TimeUnit,
@@ -226,7 +227,7 @@ class MicrogridController(Component, Controller):
                     self.power_network.connected_line.circuitbreaker.close()
                     self.power_network.connected_line.section.connect(dt, self)
                     self.failed_sections = []
-                    
+
     def check_circuitbreaker_manually(self, curr_time: Time, dt: Time):
         """
         Checks if the circuitbreaker in the distribution system is open manually.
@@ -316,20 +317,23 @@ class MicrogridController(Component, Controller):
             for sensor in sensors:
                 # If no ICT network
                 if self.ict_node is None:
-                    repair_time, line_fail_status = sensor.get_line_fail_status(dt)
+                    (
+                        repair_time,
+                        line_fail_status,
+                    ) = sensor.get_line_fail_status(dt)
                     self.sectioning_time += repair_time
                 # If both components have ICT nodes
-                elif (
-                    self.ict_node is not None
-                    and sensor.ict_node is not None
-                ):
+                elif self.ict_node is not None and sensor.ict_node is not None:
                     # If the ICT nodes are connected to each other
                     if is_connected(
                         node_1=self.ict_node,
                         node_2=sensor.ict_node,
                         network=self.ict_network,
                     ):
-                        repair_time, line_fail_status = sensor.get_line_fail_status(dt)
+                        (
+                            repair_time,
+                            line_fail_status,
+                        ) = sensor.get_line_fail_status(dt)
                         self.sectioning_time += repair_time
                     # If the ICT nodes are not connected to each other
                     else:
@@ -355,20 +359,23 @@ class MicrogridController(Component, Controller):
             for sensor in sensors:
                 # If no ICT network
                 if self.ict_node is None:
-                    repair_time, line_fail_status = sensor.get_line_fail_status(dt)
+                    (
+                        repair_time,
+                        line_fail_status,
+                    ) = sensor.get_line_fail_status(dt)
                     self.sectioning_time += repair_time
                 # If both components have ICT nodes
-                elif (
-                    self.ict_node is not None
-                    and sensor.ict_node is not None
-                ):
+                elif self.ict_node is not None and sensor.ict_node is not None:
                     # If the ICT nodes are connected to each other
                     if is_connected(
                         node_1=self.ict_node,
                         node_2=sensor.ict_node,
                         network=self.ict_network,
                     ):
-                        repair_time, line_fail_status = sensor.get_line_fail_status(dt)
+                        (
+                            repair_time,
+                            line_fail_status,
+                        ) = sensor.get_line_fail_status(dt)
                         self.sectioning_time += repair_time
                     # If the ICT nodes are not connected to each other
                     else:
