+++ docs/source/usage.rst
@@ -196,25 +196,74 @@ Creating lines::
 
 Creating circuit breaker::
 
-    E1 = CircuitBreaker(name="E1", line=L1)
+    E1 = CircuitBreaker(
+        name="E1",
+        line=L1,
+        )
 
 Creating disconnectors, here we add several disconnectors for each line. Lines can contain from 0 up to 2 disconnectors alone. If a circuit breaker is placed on a line, the line may contain 3 disconnectors:: 
 
-    DL1a = Disconnector(name="L1a", line=L1, bus=B1, circuitbreaker=E1)
-    DL1b = Disconnector(name="L1b", line=L1, bus=B2, circuitbreaker=E1)
-    DL1c = Disconnector(name="L1c", line=L1, bus=B2)
-    DL2a = Disconnector(name="L2a", line=L2, bus=B2)
-    DL2b = Disconnector(name="L2b", line=L2, bus=B3)
-    DL3a = Disconnector(name="L3a", line=L3, bus=B3)
-    DL3b = Disconnector(name="L3b", line=L3, bus=B4)
-    DL4a = Disconnector(name="L4a", line=L4, bus=B4)
-    DL4b = Disconnector(name="L4b", line=L4, bus=B5)
-    DL5a = Disconnector(name="L5a", line=L5, bus=B3)
-    DL5b = Disconnector(name="L5b", line=L5, bus=B6)
+    DL1a = Disconnector(
+        name="L1a", 
+        line=L1, bus=B1,
+        )
+    DL1b = Disconnector(
+        name="L1b",
+        line=L1,
+        bus=B2,
+        )
+    DL2a = Disconnector(
+        name="L2a",
+        line=L2,
+        bus=B2,
+        )
+    DL2b = Disconnector(
+        name="L2b",
+        line=L2,
+        bus=B3,
+        )
+    DL3a = Disconnector(
+        name="L3a",
+        line=L3,
+        bus=B3,
+        )
+    DL3b = Disconnector(
+        name="L3b",
+        line=L3,
+        bus=B4,
+        )
+    DL4a = Disconnector(
+        name="L4a",
+        line=L4,
+        bus=B4,
+        )
+    DL4b = Disconnector(
+        name="L4b",
+        line=L4,
+        bus=B5,
+        )
+    DL5a = Disconnector(
+        name="L5a",
+        line=L5,
+        bus=B3,
+        )
+    DL5b = Disconnector(
+        name="L5b",
+        line=L5,
+        bus=B6,
+        )
 
     # For backup line
-    DL6a = Disconnector(name="L6a", line=L6, bus=B4)
-    DL6b = Disconnector(name="L6b", line=L6, bus=B6)
+    DL6a = Disconnector(
+        name="L6a",
+        line=L6,
+        bus=B4,
+        )
+    DL6b = Disconnector(
+        name="L6b",
+        line=L6,
+        bus=B6,
+        )
 
 """"""""""""""""""""""""
 Add networks
@@ -223,11 +272,17 @@ Add networks
 After creating the components in the network, the components need to be added to their associated networks and the associated networks must be added to the power system. 
 First, the bus connecting to the overlying network (often transmission network) is added. In this case the overlying network is a transmission network, which is created by::
     
-    tn = Transmission(parent_network=ps, trafo_bus=B1)
+    tn = Transmission(
+        parent_network=ps,
+        trafo_bus=B1,
+        )
 
 The distribution network contains the rest of the components, and links to the transmission network with line L1. This is done by the following code snippet:: 
 
-    dn = Distribution(parent_network=tn, connected_line=L1)
+    dn = Distribution(
+        parent_network=tn,
+        connected_line=L1,
+        )
[…]
