### 2022-08-05 a434a75
updated and cleaned plotting functions
