+++ relsad/test_networks/CINELDI.py
@@ -806,7 +806,7 @@ def initialize_network(
             capacity=6,
         )
 
-    E1 = CircuitBreaker("E1", L1)
+    CircuitBreaker("E1", L1)
 
     DL1a = Disconnector("L1a", L1, B1)
     DL1b = Disconnector("L1b", L1, B2)
@@ -1018,7 +1018,7 @@ def initialize_network(
             repair_time_dist=line_stat_dist,
         )
 
-        E2 = CircuitBreaker(name="E2", line=L7)
+        CircuitBreaker(name="E2", line=L7)
 
         DL7a = Disconnector(
             name="L7a",
