### 2022-08-07 e344497
added condition for energy shedding
