### 2022-08-10 #4516 @robinroche
@stinefm @fraukewiese  Here are my comments and suggestions at this point:
- github description: adding more info on what it can do, who it can be useful for, and examples of use cases would be beneficial. The list of features could also be more specific. Mention that examples are included. Be also more specific on how to contribute, report issues or seek help.
- Paper: clarify the terms reliability and ICT, which are vague in my opinion. List examples of non traditional components and technology in the first paragraph of summary. Provide examples of metrics. Mention who is the target audience and what  other comparable packages/software could be considered and how your submission compares. In the figure, why have a datacenter? Is the communication network modeled to consider interdependence between networks?
- Installation: the installation seems to run correctly, but I could not test the package so far (see error below). It does not seem that an automatic check of dependencies is included (I could be wrong). Tests seem to be there in the code, but are not mentioned during installation.
- Examples provided in the code: I get the following error when running the 2 IEEE networks (I did not try the others): "ModuleNotFoundError: No module named 'relsad.reliability'". I am not sure why I get this, I may do something wrong.
- Documentation: the usage part focuses on the modeling but not on how this can be used to evaluate reliability. For example, how is the calculation of reliability metrics done? The creation of ICT networks is missing. Adding an example to show the interest of adding ICT could be interesting. In general, a detailed, complete, example would be nice. I would rename the part on "theory", which is vague, and put the metrics calculation on at the top level. Controllers are mentioned but I did not find information on how they work or are defined. Some API items are not described.
