### 2022-08-11 76c200f
modified CINELDI example
