### 2022-08-14 0e1490a
added run_sequential and some tests
