+++ relsad/network/components/Disconnector.py
@@ -139,6 +139,7 @@ class Disconnector(Component, Switch):
 
         ## History
         self.history = {}
+        self.initialize_history()
 
     def __str__(self):
         return self.name
