+++ relsad/network/components/CircuitBreaker.py
@@ -122,6 +122,7 @@ class CircuitBreaker(Component, Switch):
 
         ## History
         self.history = {}
+        self.initialize_history()
 
     def __str__(self):
         return self.name
