+++ relsad/network/components/MicrogridController.py
@@ -312,10 +312,7 @@ class MicrogridController(Component, Controller):
         ]
         for section in disconnected_sections:
             sensors = unique(
-                [
-                    switch.line.sensor
-                    for switch in section.switches
-                ]
+                [switch.line.sensor for switch in section.switches]
             )
             num_fails = 0
             need_manual_attention = False
@@ -358,10 +355,7 @@ class MicrogridController(Component, Controller):
         # Loop connected sections
         for section in connected_sections:
             sensors = unique(
-                [
-                    switch.line.sensor
-                    for switch in section.switches
-                ]
+                [switch.line.sensor for switch in section.switches]
             )
             num_fails = 0
             need_manual_attention = False
