+++ relsad/network/components/EVPark.py
@@ -200,6 +200,7 @@ class EVPark(Component):
 
         ## History
         self.history = {}
+        self.initialize_history()
 
     def __str__(self):
         return self.name
