+++ relsad/network/components/Sensor.py
@@ -163,6 +163,7 @@ class Sensor(Component):
         ## History
         self.history = {}
         self.monte_carlo_history = {}
+        self.initialize_history()
 
     def __str__(self):
         return self.name

[…]
