### 2022-08-25 56e20c4
added cd, fixed bug and moved examples
