+++ examples/scripts/IEEE33/load_and_gen_data.py
@@ -1,331 +1,0 @@
-import numpy as np
-import pandas as pd
-import matplotlib.pyplot as plt
-import os
-
-
-def weather_generation_data(
-    path: str = os.path.join("data", "weather_data_rygge.csv"),
-):
-
-    # Generate temperatures for a whole year
-    data = pd.read_csv(path, header=None, skiprows=1)
-
-    temp_profiles = data[1].values.reshape(365, 24)
-    wind_profiles = data[2].values.reshape(365, 24)
-    solar_profiles = data[3].values.reshape(365, 24)
-
-    return temp_profiles, wind_profiles, solar_profiles
-
-
-def load_data(
-    temp: np.array,
-    path: str = os.path.join("data", "load_profiles_fasit.csv"),
-):
-
-    # Generating load profiles:
-
-    data = pd.read_csv(
-        path,
-        sep=";",
-        decimal=",",
-        header=None,
-    ).transpose()
-    data = data.stack().str.replace(",", ".").unstack()
-
-    house_A = data[10].values[4:-1].astype(float)
-    house_B = data[11].values[4:-1].astype(float)
-
-    farm_A = data[2].values[4:-1].astype(float)
-    farm_B = data[3].values[4:-1].astype(float)
-
-    industry2_A = data[26].values[4:-1].astype(float)
-    industry2_B = data[27].values[4:-1].astype(float)
-
-    trade_A = data[42].values[4:-1].astype(float)
-    trade_B = data[43].values[4:-1].astype(float)
-
-    office_A = data[50].values[4:-1].astype(float)
-    office_B = data[51].values[4:-1].astype(float)
-
-    load_house = np.zeros_like(temp)
-    load_farm = np.zeros_like(temp)
-    load_microgrid = np.zeros_like(temp)
-    load_industry2 = np.zeros_like(temp)
-    load_trade = np.zeros_like(temp)
-    load_office = np.zeros_like(temp)
-
-    for i in range(len(temp)):
-        for j in range(len(temp[i])):
-            if temp[i, j] > 20:
-                temp[i, j] = 20
-
-            load_house[i, j] = house_A[j] * temp[i, j] + house_B[j]
-            load_farm[i, j] = farm_A[j] * temp[i, j] + farm_B[j]
-            load_microgrid[i, j] = farm_A[j] * temp[i, j] + farm_B[j]
-            load_industry2[i, j] = industry2_A[j] * temp[i, j] + industry2_B[j]
-            load_trade[i, j] = trade_A[j] * temp[i, j] + trade_B[j]
-            load_office[i, j] = office_A[j] * temp[i, j] + office_B[j]
-
-    return (
-        load_house.reshape((8760)) * 1e-3,
-        load_farm.reshape((8760)) * 1e-3,
-        load_microgrid.reshape((8760)) * 1e-3,
-        load_industry2.reshape((8760)) * 1e-3,
-        load_trade.reshape((8760)) * 1e-3,
-        load_office.reshape((8760)) * 1e-3,
-    )
-
-
-def wind_power(
-    wind: np.array,
-):
-
-    # Wind power
-
-    # https://en.wind-turbine-models.com/turbines/1829-aeolia-windtech-d2cf-200
-
-    N = 5
-    alpha = 0.25
-    H_0 = 2
-    H = 50
-
-    v = np.zeros_like(wind)
-    P_out = np.zeros_like(wind)
-
-    P_rated = np.array(
-        [
-            0,
-            0,
-            0,
-            34.8,
-            69.7,
-            123.5,
-            200.7,
-            306.6,
-            446.4,
-            495.0,
-            500,
-            500,
-            500,
[…]
