### 2022-08-25 802cd3c
added bug fix for MicrogridController, shortened example test durations
