+++ tests/test_examples.py
@@ -1,4 +1,5 @@
 import os
+from relsad.Time import TimeStamp
 from relsad.examples.CINELDI.simulation import run_simulation as run_cineldi
 from relsad.examples.IEEE33.simulation import run_simulation as run_ieee33
 from relsad.examples.IEEE16_modified.simulation import run_simulation as run_ieee16_modified
@@ -14,6 +15,22 @@ def test_run_cineldi():
             "load",
             "data",
         ),
+        start_time=TimeStamp(
+            year=2019,
+            month=1,
+            day=1,
+            hour=0,
+            minute=0,
+            second=0,
+        ),
+        stop_time=TimeStamp(
+            year=2019,
+            month=1,
+            day=3,
+            hour=0,
+            minute=0,
+            second=0,
+        ),
         save_flag=False,
     )
 
@@ -25,6 +42,22 @@ def test_run_ieee16_modified():
             "examples",
             "load",
             "data",
+        ),
+        start_time=TimeStamp(
+            year=2019,
+            month=1,
+            day=1,
+            hour=0,
+            minute=0,
+            second=0,
+        ),
+        stop_time=TimeStamp(
+            year=2019,
+            month=1,
+            day=3,
+            hour=0,
+            minute=0,
+            second=0,
         ),
         save_flag=False,
     )
@@ -38,12 +71,44 @@ def test_run_ieee33():
             "load",
             "data",
         ),
+        start_time=TimeStamp(
+            year=2019,
+            month=1,
+            day=1,
+            hour=0,
+            minute=0,
+            second=0,
+        ),
+        stop_time=TimeStamp(
+            year=2019,
+            month=1,
+            day=3,
+            hour=0,
+            minute=0,
+            second=0,
+        ),
         save_flag=False,
     )
 
 
 def test_run_rbts2():
     run_rbts2(
+        start_time=TimeStamp(
+            year=2019,
+            month=1,
+            day=1,
+            hour=0,
+            minute=0,
+            second=0,
+        ),
+        stop_time=TimeStamp(
+            year=2019,
+            month=1,
+            day=3,
+            hour=0,
+            minute=0,
+            second=0,
+        ),
         save_flag=False,
     )
 
@@ -55,6 +120,22 @@ def test_run_test10():
             "examples",
             "load",
             "data",
+        ),
+        start_time=TimeStamp(
+            year=2019,
+            month=1,
+            day=1,
+            hour=0,
+            minute=0,
+            second=0,
+        ),
+        stop_time=TimeStamp(
+            year=2019,
+            month=1,
+            day=3,
+            hour=0,
+            minute=0,
+            second=0,
         ),
         save_flag=False,
     )
