+++ relsad/examples/IEEE33/simulation.py
@@ -43,6 +43,22 @@ def run_simulation(
     battery_capacity: float = 1,  # MWh
     random_seed: int = 2837314,
     iterations: int = 2,
+    start_time: TimeStamp = TimeStamp(
+        year=2019,
+        month=1,
+        day=1,
+        hour=0,
+        minute=0,
+        second=0,
+    ),
+    stop_time: TimeStamp = TimeStamp(
+        year=2020,
+        month=1,
+        day=1,
+        hour=0,
+        minute=0,
+        second=0,
+    ),
     save_iterations: list = [1, 2],
     data_dir: str = os.path.join(
         os.pardir,
@@ -84,22 +100,8 @@ def run_simulation(
 
     sim.run_monte_carlo(
         iterations=iterations,
-        start_time=TimeStamp(
-            year=2019,
-            month=1,
-            day=1,
-            hour=0,
-            minute=0,
-            second=0,
-        ),
-        stop_time=TimeStamp(
-            year=2020,
-            month=1,
-            day=1,
-            hour=0,
-            minute=0,
-            second=0,
-        ),
+        start_time=start_time,
+        stop_time=stop_time,
         time_step=Time(5, TimeUnit.MINUTE),
         time_unit=TimeUnit.HOUR,
         save_iterations=save_iterations,
