+++ .github/workflows/ci-cd.yml
@@ -58,14 +58,6 @@ jobs:
     - name: Install package
       run: poetry install
 
-    - name: Use Python Semantic Release to prepare release
-      env:
-        GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-      run: |
-          git config user.name github-actions
-          git config user.email github-actions@github.com
-          poetry run semantic-release publish
-
     - name: Publish to TestPyPI
       uses: pypa/gh-action-pypi-publish@release/v1
       with:
