+++ .github/workflows/ci-cd.yml
@@ -55,8 +55,8 @@ jobs:
     - name: Install poetry
       uses: snok/install-poetry@v1
 
-    - name: Install package
-      run: poetry install
+    - name: Build package
+      run: poetry build
 
     - name: Publish to TestPyPI
       uses: pypa/gh-action-pypi-publish@release/v1
@@ -70,7 +70,7 @@ jobs:
           pip install \
           --index-url https://test.pypi.org/simple/ \
           --extra-index-url https://pypi.org/simple \
-          pycounts
+          relsad
 
     - name: Publish to PyPI
       uses: pypa/gh-action-pypi-publish@release/v1
