### 2022-08-25 fbe40a2
updated pyproject.toml with more information
