+++ .pre-commit-config.yaml
@@ -1,26 +1,24 @@
 repos:
--   repo: https://github.com/ambv/black
-    rev: stable
+  - repo: https://github.com/ambv/black
+    rev: 22.6.0
     hooks:
     - id: black
       language_version: python3
--   repo: https://gitlab.com/pycqa/flake8
-    rev: 3.7.5
+  - repo: https://gitlab.com/pycqa/flake8
+    rev: 3.9.2
     hooks:
     - id: flake8
--   repo: git://github.com/FalconSocial/pre-commit-python-sorter
-    sha: 1.0.4
+  - repo: https://github.com/pycqa/isort
+    rev: 5.10.1
     hooks:
-    - id: python-import-sorter
-      args: ['--silent-overwrite']
--   repo: https://github.com/asottile/blacken-docs
+      - id: isort
+        args: ["--line-length", "79"]
+  - repo: https://github.com/asottile/blacken-docs
     rev: v1.12.1
     hooks:
     -   id: blacken-docs
-        additional_dependencies: [black==...]
--   repo: https://github.com/pre-commit/pre-commit-hooks
+  - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v4.3.0
     hooks:
     -   id: trailing-whitespace
-    -   id: requirements-txt-fixer
     -   id: end-of-file-fixer
