### 2022-08-26 d82404d
fixed pre-commit config, updated docs
