+++ relsad/results/storage.py
@@ -1,5 +1,6 @@
-import pandas as pd
 import os
+
+import pandas as pd
 
 
 def save_history(comp_list: list, attribute: str, save_dir: str):
