+++ tests/test_relsad.py
@@ -2,4 +2,4 @@ from relsad import __version__
 
 
 def test_version():
-    assert __version__ == "0.1.0"
+    assert __version__ == "0.0.1"
