+++ .github/workflows/ci-cd.yml
@@ -33,7 +33,7 @@ jobs:
   cd:
     # Only run this job if the "ci" job passes
     needs: ci
-    
+
     # Only run this job if new work is pushed to "main"
     if: github.event_name == 'push' && github.ref == 'refs/heads/main'
 
@@ -55,8 +55,21 @@ jobs:
     - name: Install poetry
       uses: snok/install-poetry@v1
 
-    - name: Build package
+    - name: build package
       run: poetry build
+
+    - name: rename package
+      run: python -c "import datetime; \
+                       import glob; \
+                       import os; \
+                       file = glob.glob('dist/*.whl')[0]; \
+                       new_file = file.replace( \
+                          '.whl', \
+                          '_'+str( \
+                              datetime.datetime.now() \
+                              .strftime('%Y%m%d_%H-%M-%S') \
+                          )+'.whl'); \
+                        os.rename(file, new_file)"
 
     - name: Publish to TestPyPI
       uses: pypa/gh-action-pypi-publish@release/v1
