+++ .github/workflows/ci-cd.yml
@@ -59,19 +59,19 @@ jobs:
       run: poetry build
 
     - name: rename package
-      run: >
-        python -c
-        "import datetime;
-         import glob;
-         import os;
-         file = glob.glob('dist/*.whl')[0];
-         new_file = file.replace(
-            '.whl',
-            '_'+str(
-                datetime.datetime.now()
-                .strftime('%Y%m%d_%H-%M-%S')
-            )+'.whl');
-         os.rename(file, new_file)"
+      run: |
+        python -c \
+        "import datetime; \
+        import glob; \
+        import os; \
+        file = glob.glob('dist/*.whl')[0]; \
+        new_file = file.replace( \
+        '.whl', \
+        '_'+str( \
+        datetime.datetime.now() \
+        .strftime('%Y%m%d_%H-%M-%S') \
+        )+'.whl'); \
+        os.rename(file, new_file)" \
 
     - name: Publish to TestPyPI
       uses: pypa/gh-action-pypi-publish@release/v1
