+++ .github/workflows/ci-cd.yml
@@ -69,7 +69,7 @@ jobs:
         'py3', \
         'py3-'+str( \
         datetime.datetime.now() \
-        .strftime('%Y%m%d_%H-%M-%S') \
+        .strftime('%Y%m%d_%H%M%S') \
         )); \
         os.rename(file, new_file)" \
 
