### 2022-08-26 44d6d77
updated version to 0.0.2
