+++ .github/workflows/ci-cd.yml
@@ -58,21 +58,6 @@ jobs:
     - name: build package
       run: poetry build
 
-    - name: rename package
-      run: |
-        python -c \
-        "import datetime; \
-        import glob; \
-        import os; \
-        file = glob.glob('dist/*.whl')[0]; \
-        new_file = file.replace( \
-        'py3', \
-        'py3-'+str( \
-        datetime.datetime.now() \
-        .strftime('%Y%m%d_%H%M%S') \
-        )); \
-        os.rename(file, new_file)" \
-
     - name: Publish to TestPyPI
       uses: pypa/gh-action-pypi-publish@release/v1
       with:
