+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "relsad"
-version = "0.0.1"
+version = "0.0.2"
 description = "A package that facilitates reliability investigations in power systems"
 readme = "README.rst"
 repository = "https://github.com/stinefm/relsad"
