### 2022-08-30 8c1dc67
feat: add CI/CD workflow
