+++ joss/paper.md
@@ -37,25 +37,27 @@ bibliography: paper.bib
 
 # Summary
 
-With increased penetration of new technology in the electrical power systems such as Renewable Energy Resources, flexible resources, and Information and Communication Technology (ICT), the power systems become more complex and with more dependencies. 
-The traditional reliability analysis methods for power systems do not consider the new components and technology, therefore, new considerations need to be taken to address these new changes in the system. 
+The electrical power distribution system is under constant development and multiple changes will occur in the upcoming years. Through the integration of new technology such as Renewable Energy Resources (RES) and flexible resources such as microgrids, battery energy storage systems, and electrical vehicles, the distribution system becomes more active where bidirectional power flow is possible. In an active distribution system, utilization and control of the different sources are possible. Additionally, the system becomes smarter through the integration of Information and Communication Technology (ICT) where intelligent monitoring, automated control, and communication are possible. These changes increase the complexity of the distribution network and new dependencies and interdependencies in the system arise with the modernization of the system. 
+
+The traditional reliability analysis methods for power systems do not consider the new components and technology, and the behavior and impact these have on the distribution network. Therefore, new considerations need to be taken to address these new changes in the system. 
 This paper presents an open-source reliability assessment tool for smart and active distribution systems named `RELSAD`. 
 
-`RELSAD` -- RELiability tool for Smart and Active Distribution networks, is a Python-based reliability assessment tool that aims to function as a foundation for reliability calculation of modern distribution systems. 
+`RELSAD` -- RELiability tool for Smart and Active Distribution networks is a Python-based reliability assessment tool that aims to function as a foundation for reliability calculation of modern distribution systems. RELSAD can be used by scientists, engineers, and Distribution System Operators (DSO) for the planning and operation of modern distribution system where the goal is to investigate the reliability of a given system.
 <!--`RELSAD` is a Python-based package tool that aims to function as a foundation for reliability calculation for smart and active distribution systems. -->
-The tool uses Monte Carlo simulation and stochastic variation to simulate the reliability of a distribution system. The package supports user-selected time increment steps over a user-defined time period. In the tool, active components such as microgrids, wind power, solar power, batteries, and electrical vehicles are implemented. To evaluate smart power systems, ICT components such as automated switches, sensors, and control system for the power grid are also implemented. In addition to component implementation, in order to evaluate the reliability of such complex systems, the complexity, dependencies within a system, and interdependencies between systems and components are accounted for. 
+The tool allows for Monte Carlo simulation based reliability analysis of modern distribution networks, and sequential simulation of the network behavior with user-defined loading and failure evolution to investigate the impact of the introduction of for instance ICT components. The package supports user-selected time increment steps over a user-defined time period. In the tool, active components such as microgrids, distributed generation, batteries, and electrical vehicles are implemented. To evaluate smart power systems, ICT components such as automated switches, sensors, and control system for the power grid are also implemented. In addition to component implementation, in order to evaluate the reliability of such complex systems, the complexity, dependencies within a system, and interdependencies between systems and components are accounted for. 
+
 
 <!--- the dependencies between the different components and networks are included. --->
 
 `RELSAD` offers the following features:
 
-* A foundation for calculating the reliability of modern distribution systems with smart and active components that account for the increased complexity in the power system. 
+* A foundation for calculating the reliability of modern distribution systems with smart and active components that account for the increased complexity of the system. 
 
 * An extensive reliability analysis foundation with important and new reliability indices that aims to give a complete picture of the reliability of both the power system and the components in the system.
 
[…]
