+++ relsad/visualization/plotting/topology.py
@@ -1,0 +1,479 @@
+import os
+
+import matplotlib.pyplot as plt
+import numpy as np
+import pandas as pd
+
+from relsad.network.components import (
+    Bus,
+    CircuitBreaker,
+    Disconnector,
+    EVPark,
+    IntelligentSwitch,
+    Line,
+    Sensor,
+)
+
+from relsad.Time import TimeUnit
+
+
+def plot_topology(
+    buses: list,
+    lines: list,
+    bus_text: bool = True,
+    line_text: bool = False,
+    circuitbreaker_text: bool = False,
+    disconnector_text: bool = False,
+    intelligent_switch_text: bool = False,
+    sensor_text: bool = False,
+    text_dx: tuple = (0, -0.1),
+    **kwargs
+):
+
+    """
+    Plots the system topology
+
+    Parameters
+    ----------
+    buses : list
+        List with Bus elements in the system
+    lines : list
+        List with Line elements in the system
+    bus_text : bool
+        Flag determining if bus name will be plotted
+    line_text : bool
+        Flag determining if line name will be plotted
+    circuitbreaker_text : bool
+        Flag determining if circuitbreaker name will be plotted
+    disconnector_text : bool
+        Flag determining if disconnector name will be plotted
+    intelligent_switch_text : bool
+        Flag determining if intelligent switch name will be plotted
+    sensor_text : bool
+        Flag determining if sensor name will be plotted
+    **kwargs : dict
+        Plotting keyword arguments.
+
+    Returns
+    ----------
+    fig : figure
+        Figure of the system topology
+    None
+
+    """
+    # Handle keyword arguments
+
+    left = kwargs["left"] if "left" in kwargs else 0.02
+    right = kwargs["right"] if "right" in kwargs else 0.98
+    bottom = kwargs["bottom"] if "bottom" in kwargs else 0.1
+    top = kwargs["top"] if "top" in kwargs else 0.75
+
+    text_size = kwargs["text_size"] if "text_size" in kwargs else 6
+
+    ncol = kwargs["ncol"] if "ncol" in kwargs else None
+
+    # Clear used keyword arguments
+    kwargs.pop("left", None)
+    kwargs.pop("right", None)
+    kwargs.pop("bottom", None)
+    kwargs.pop("top", None)
+    kwargs.pop("text_size", None)
+    kwargs.pop("ncol", None)
+
+    fig, ax = plt.subplots(**kwargs)
+
+    fig.subplots_adjust(
+        left=left,
+        bottom=bottom,
+        right=right,
+        top=top,
+        wspace=None,
+        hspace=None,
+    )
+    legends = {}
+    for bus in buses:
+        _plot_bus(
+            ax=ax,
+            bus=bus,
+            text=bus_text,
+            text_size=text_size,
+            text_dx=text_dx,
+        )
+        legends["Bus"] = Bus.handle
+        if bus.ev_park is not None:
+            legends["EV park"] = EVPark.handle
+    for line in lines:
+        _plot_line(
+            ax=ax,
+            line=line,
+            text=line_text,
+            text_size=text_size,
+        )
+        legends["Line"] = Line.handle
+        if line.circuitbreaker is not None:
+            _plot_circuitbreaker(
+                ax=ax,
+                line=line,
+                text=circuitbreaker_text,
+                text_size=text_size,
+            )
+            legends["Circuit breaker"] = CircuitBreaker.handle
+        for discon in line.disconnectors:
+            _plot_disconnector(
+                ax=ax,
+                discon=discon,
+                text=disconnector_text,
+                text_size=text_size,
+            )
+            legends["Disconnector"] = Disconnector.handle
[…]
