+++ relsad/network/systems/__init__.py
@@ -15,3 +15,8 @@ from .PowerNetwork import PowerNetwork
 from .PowerSystem import PowerSystem
 from .SubSystem import SubSystem
 from .Transmission import Transmission
+
+__all__ = []
+for v in dir():
+    if not v.startswith("__") and v != "relsad":
+        __all__.append(v)
