+++ relsad/network/components/MainController.py
@@ -433,13 +433,14 @@ class MainController(Component, Controller):
 
         """
         if save_flag:
+            time = curr_time.get_unit_quantity(curr_time.unit)
             self.history["sectioning_time"][
-                curr_time
+                time
             ] = self.sectioning_time.get_unit_quantity(curr_time.unit)
             self.history["remaining_repair_time"][
-                curr_time
+                time
             ] = self.remaining_repair_time.get_unit_quantity(curr_time.unit)
-            self.history["state"][curr_time] = self.state.value
+            self.history["state"][time] = self.state.value
 
     def get_history(self, attribute: str):
         """

[…]
