+++ .github/workflows/ci-cd.yml
@@ -60,7 +60,7 @@ jobs:
     
     - name: Use Python Semantic Release to prepare release
       env:
-        GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        GH_TOKEN: ${{ secrets.SEMANTIC_RELEASE_TOKEN }}
       run: |
           git config user.name github-actions
           git config user.email github-actions@github.com
