### 2022-08-30 #4516 @stinefm
Hi @robinroche and @fraukewiese!

Thank you for the comments! I have been working on improvements based on your feedback. I will sum them up with bullet points responding to your comments:

•	Github description: I have updated the info on what the tool can do with examples of use cases. The list of features is updated and should be more specific now. I have now mentioned that examples are included, how to contribute, report issues, and seek help.

•	Paper: I have tried to clarify the terms reliability and ICT, and hope I am able to illustrate what I mean by the terms. I have added a list of examples for the technologies in the summary. Regarding metrics, I have included a paragraph related to reliability indices. The potential users are specified. Regarding packages/software, I have not found any similar software that are open available, and by that not compared any software. I have extended the description related to the structure of RELSAD, hopefully it is clearer now.

•	Installation: The issues regarding the installation should be fixed now. I have created an automatic upload to PyPI, so that the package can be installed with pip directly. I have added info about how to test the package in the documentation.

•	Documentation: I have expanded the usage part of the documentation to include a tutorial of the basic features of RELSAD and how to run an example and extract reliability indices/results and obtain statistical metrics. I have added examples of how ICT networks can be created in the tutorial. I agree, the use of the term theory is vague and is now changed. For the controllers there is a brief explanation in the background page, and I am considering whether to add more information about the implementation. The API should be updated to include all the items.

I hope you agree with the changes.

Best
Stine
