+++ CHANGELOG.rst
@@ -2,3 +2,9 @@
 Changelog
 =========
 
+<!--next-version-placeholder-->
+
+v0.1.0 (30/08/2022)
+-------------------
+
+- First release of `relsad`
