+++ relsad/examples/CINELDI/network.py
@@ -142,6 +142,7 @@ def initialize_network(
         # Controller
         ICTNController = ICTNode(
             name="ICTNController",
+            coordinate=[0, 2],
         )
         # Sensors
         ICTNSL1 = ICTNode(
@@ -237,27 +238,35 @@ def initialize_network(
         ## Access nodes
         ICTNA1 = ICTNode(
             name="ICTNA1",
+            coordinate=[0.5, 2.25],
         )
         ICTNA2 = ICTNode(
             name="ICTNA2",
+            coordinate=[0.5, 2.0],
         )
         ICTNA3 = ICTNode(
             name="ICTNA3",
+            coordinate=[0.75, 2.0],
         )
         ICTNA4 = ICTNode(
             name="ICTNA4",
+            coordinate=[0.75, 2.25],
         )
         ICTNA5 = ICTNode(
             name="ICTNA5",
+            coordinate=[2.25, 1.5],
         )
         ICTNA6 = ICTNode(
             name="ICTNA6",
+            coordinate=[2.5, 1.5],
         )
         ICTNA7 = ICTNode(
             name="ICTNA7",
+            coordinate=[2.5, 1.75],
         )
         ICTNA8 = ICTNode(
             name="ICTNA8",
+            coordinate=[2.25, 1.75],
         )
         ict_nodes.extend(
             [
@@ -275,39 +284,51 @@ def initialize_network(
         ## Network nodes
         ICTN1 = ICTNode(
             name="ICTN1",
+            coordinate=[1.0, 1.75],
         )
         ICTN2 = ICTNode(
             name="ICTN2",
+            coordinate=[1.25, 1.5],
         )
         ICTN3 = ICTNode(
             name="ICTN3",
+            coordinate=[1.5, 1.75],
         )
         ICTN4 = ICTNode(
             name="ICTN4",
+            coordinate=[1.75, 2.0],
         )
         ICTN5 = ICTNode(
             name="ICTN5",
+            coordinate=[2.25, 1.75],
         )
         ICTN6 = ICTNode(
             name="ICTN6",
+            coordinate=[2.5, 2.0],
         )
         ICTN7 = ICTNode(
             name="ICTN7",
+            coordinate=[2.5, 2.25],
         )
         ICTN8 = ICTNode(
             name="ICTN8",
+            coordinate=[2.25, 2.5],
         )
         ICTN9 = ICTNode(
             name="ICTN9",
+            coordinate=[1.75, 2.5],
         )
         ICTN10 = ICTNode(
             name="ICTN10",
+            coordinate=[1.5, 2.0],
         )
         ICTN11 = ICTNode(
             name="ICTN11",
+            coordinate=[1.5, 2.5],
         )
         ICTN12 = ICTNode(
             name="ICTN12",
+            coordinate=[1.25, 2.25],
         )
         ict_nodes.extend(
             [
@@ -391,48 +412,55 @@ def initialize_network(
         )
         ICTL9 = ICTLine(
             name="ICTL9",
-            fnode=ICTNISW2b,
-            tnode=ICTNA5,
+            fnode=ICTNISW2a,
+            tnode=ICTNA2,
             repair_time_dist=ict_line_repair_time_stat_dist,
             fail_rate_per_year=fail_rate_ict_line,
         )
         ICTL10 = ICTLine(
             name="ICTL10",
-            fnode=ICTNISW3a,
+            fnode=ICTNISW2b,
             tnode=ICTNA5,
             repair_time_dist=ict_line_repair_time_stat_dist,
             fail_rate_per_year=fail_rate_ict_line,
         )
         ICTL11 = ICTLine(
             name="ICTL11",
-            fnode=ICTNISW3b,
+            fnode=ICTNISW3a,
             tnode=ICTNA5,
[…]
