### 2022-09-04 2a74565
feat: added callback to run_monte_carlo and indices to sequence history
