+++ relsad/simulation/monte_carlo/history.py
@@ -2,72 +2,7 @@ import copy
 import os
 
 from relsad.network.systems import PowerSystem
-from relsad.reliability.indices import (
-    ASAI,
-    ASUI,
-    CAIDI,
-    ENS,
-    SAIDI,
-    SAIFI,
-    EV_Duration,
-    EV_Index,
-    EV_Interruption,
-)
 from relsad.results.storage import save_monte_carlo_history_from_dict
-from relsad.Time import Time
-from relsad.visualization.plotting import plot_monte_carlo_history
-
-
-def plot_network_monte_carlo_history(power_system: PowerSystem, save_dir: str):
-    """
-    Plots the history of the energy.shedding in the power system
-
-    Parameters
-    ----------
-    power_system : PowerSystem
-        A power system element
-    save_dir : str
-        The saving path
-
-    Returns
-    ----------
-    None
-
-    """
-    network_state_list = [
-        "acc_p_energy_shed",
-        "acc_q_energy_shed",
-        "SAIFI",
-        "SAIDI",
-        "CAIDI",
-        "ASAI",
-        "ASUI",
-        "ENS",
-        "EV_Index",
-        "EV_Interruption",
-        "EV_Duration",
-    ]
-    power_system_save_dir = os.path.join(save_dir, power_system.name)
-    for state_var in network_state_list:
-        plot_monte_carlo_history(
-            [power_system], state_var, power_system_save_dir
-        )
-    for network in power_system.child_network_list:
-        network_save_dir = os.path.join(save_dir, network.name)
-        for state_var in network_state_list:
-            plot_monte_carlo_history([network], state_var, network_save_dir)
-    bus_state_list = [
-        "acc_p_energy_shed",
-        "acc_q_energy_shed",
-        "avg_outage_time",
-        "acc_outage_time",
-        "interruption_fraction",
-        "acc_interruptions",
-    ]
-    for bus in power_system.buses:
-        bus_save_dir = os.path.join(save_dir, bus.name)
-        for state_var in bus_state_list:
-            plot_monte_carlo_history([bus], state_var, bus_save_dir)
 
 
 def save_network_monte_carlo_history(
@@ -94,19 +29,7 @@ def save_network_monte_carlo_history(
     """
     if not os.path.isdir(save_dir):
         os.makedirs(save_dir)
-    network_state_list = [
-        "acc_p_energy_shed",
-        "acc_q_energy_shed",
-        "SAIFI",
-        "SAIDI",
-        "CAIDI",
-        "ASAI",
-        "ASUI",
-        "ENS",
-        "EV_Index",
-        "EV_Interruption",
-        "EV_Duration",
-    ]
+    network_state_list = list(save_dict[power_system.name].keys())
     power_system_save_dir = os.path.join(save_dir, power_system.name)
     for state_var in network_state_list:
         save_monte_carlo_history_from_dict(
@@ -118,238 +41,20 @@ def save_network_monte_carlo_history(
             save_monte_carlo_history_from_dict(
                 save_dict, [network], state_var, network_save_dir
             )
-    bus_state_list = [
-        "acc_p_energy_shed",
-        "acc_q_energy_shed",
-        "avg_outage_time",
-        "acc_outage_time",
-        "interruption_fraction",
-        "acc_interruptions",
-    ]
-    ev_park_state_list = [
-        "acc_num_interruptions",
-        "acc_exp_interruptions",
-        "acc_exp_car_interruptions",
-        "acc_interruption_duration",
-        "acc_available_num_cars",
-        "num_cars",
-    ]
     for bus in power_system.buses:
         bus_save_dir = os.path.join(save_dir, bus.name)
+        bus_state_list = list(save_dict[bus.name].keys())
[…]
