### 2022-09-04 1924703
refactor: moved history initializations and updates, changed plotting features
