+++ relsad/network/systems/PowerSystem.py
@@ -1,6 +1,17 @@
 import numpy as np
 
 from relsad.network.components import Bus, ICTLine, ICTNode, Line
+from relsad.reliability.indices import (
+    ASAI,
+    ASUI,
+    CAIDI,
+    ENS,
+    SAIDI,
+    SAIFI,
+    EV_Duration,
+    EV_Index,
+    EV_Interruption,
+)
 from relsad.Time import Time, TimeStamp
 from relsad.topology.sectioning import create_sections, get_section_list
 from relsad.utils import INF, eq, unique
@@ -105,26 +116,47 @@ class PowerSystem(PowerNetwork):
     update_ev_parks(fail_duration, dt, start_time, curr_time)
         Updates the EV parks in the power system
     update_fail_status(dt)
-        Updates the failure status for each component that can fail in the power system
+        Updates the failure status for each component that
+        can fail in the power system
     get_system_load()
         Returns the system load at the current time in MW and MVar
     prepare_load_data(time_indices)
         Prepares the load data for the buses in the power system
     prepare_prod_data(time_indices)
-        Prepares the production data for the production components in the power system
+        Prepares the production data for the production components
+        in the power system
     set_load_and_cost(inc_idx)
         Sets the bus load and cost in MW based on load and cost profiles
         in the current increment for the power system
     set_prod(inc_idx)
-        Sets the generation (generation units, batteries, EV parks) at the buses in the power system
+        Sets the generation (generation units, batteries, EV parks)
+        at the buses in the power system
     failes_comp()
-        Returns True if the power system contains a failed compoent, False otherwise
+        Returns True if the power system contains a
+        failed compoent, False otherwise
     full_batteries()
-        Returns True if the batteries in the power system are full, and False otherwise
+        Returns True if the batteries in the power system
+        are full, and False otherwise
     reset_load_flow_data()
         Resets the variables used in the load flow analysis
     get_monte_carlo_history(attribute)
-        Returns the specified history variable from the Monte Carlo simulation
+        Returns the specified history variable
+        from the Monte Carlo simulation
+    initialize_sequence_history()
+        Initializes the dictionaries used for sequence history variables
+    update_sequence_history()
+        Updates the sequence history variables of the power system
+    initialize_monte_carlo_history()
+        Initializes the lists used for history variable
+        from the Monte Carlo simulation
+    update_monte_carlo_history()
+        Updates the history dictionary from the Monte Carlo simulation
+    update_monte_carlo_child_network_history()
+        Updates the history dictionary for the child networks
+        in the Monte Carlo simulation
+    update_monte_carlo_comp_history()
+        Updates the component values for the system
+        from the Monte Carlo simulation
     get_history(attribute)
         Returns the specified history variable
     reset_energy_shed_variables()
@@ -161,24 +193,24 @@ class PowerSystem(PowerNetwork):
         self.q_energy_shed = 0
         self.acc_q_energy_shed = 0
         # Sub-systems
-        self.sub_systems = list()
+        self.sub_systems: list = []
         # Components
-        self.buses = list()
-        self.ev_parks = list()
-        self.batteries = list()
-        self.productions = list()
-        self.ict_nodes = list()
-        self.lines = list()
-        self.sensors = list()
-        self.circuitbreakers = list()
-        self.disconnectors = list()
-        self.intelligent_switches = list()
-        self.ict_lines = list()
+        self.buses: list = []
+        self.ev_parks: list = []
+        self.batteries: list = []
+        self.productions: list = []
+        self.ict_nodes: list = []
+        self.lines: list = []
[…]
