+++ relsad/visualization/plotting/topology.py
@@ -1,8 +1,4 @@
-import os
-
 import matplotlib.pyplot as plt
-import numpy as np
-import pandas as pd
 
 from relsad.network.components import (
     Bus,
@@ -12,14 +8,16 @@ from relsad.network.components import (
     IntelligentSwitch,
     Line,
     Sensor,
+    ICTNode,
+    ICTLine,
 )
-
-from relsad.Time import TimeUnit
 
 
 def plot_topology(
     buses: list,
     lines: list,
+    ict_nodes: list = None,
+    ict_lines: list = None,
     bus_text: bool = True,
     line_text: bool = False,
     circuitbreaker_text: bool = False,
@@ -137,11 +135,30 @@ def plot_topology(
         if line.sensor:
             _plot_sensor(
                 ax=ax,
-                line=line,
+                sensor=line.sensor,
                 text=sensor_text,
                 text_size=text_size,
             )
             legends["Sensor"] = Sensor.handle
+    if ict_nodes is not None:
+        for ict_node in ict_nodes:
+            _plot_ict_node(
+                ax=ax,
+                ict_node=ict_node,
+                text=False,
+                text_size=text_size,
+                text_dx=text_dx,
+            )
+            legends["ICT node"] = ICTNode.handle
+    if ict_lines is not None:
+        for ict_line in ict_lines:
+            _plot_ict_line(
+                ax=ax,
+                ict_line=ict_line,
+                text=False,
+                text_size=text_size,
+            )
+            legends["ICT line"] = ICTLine.handle
 
     plt.figlegend(
         legends.values(),
@@ -159,6 +176,68 @@ def plot_topology(
     return fig
 
 
+def _plot_bus(
+    ax: plt.axis,
+    bus: Bus,
+    text: bool = False,
+    text_size: int = 8,
+    text_dx: tuple = (0, 0),
+):
+    """
+    Plot bus
+
+    Parameters
+    ----------
+    ax : matplotlib.axes.Axes
+        Plot axis
+    bus : Bus
+        Bus element
+    text : bool
+        Flag determining if bus name will be plotted
+    text_size : int
+        The size of the text in the plot
+    text_dx : tuple
+        The distance from the text to the center point
+
+    Returns
+    ----------
+    None
+
+    """
+    x, y = bus.coordinate
+    ax.plot(
+        x,
+        y,
+        marker=bus.marker,
+        markeredgewidth=bus.handle.get_markeredgewidth(),
+        markersize=bus.size,
+        linestyle="None",
+        color=bus.color,
+        clip_on=False,
+        zorder=3,
+    )
+    if text:
+        ax.text(
+            x + text_dx[0],
+            y + text_dx[1],
+            bus.name,
+            ha="center",
+            va="center",
+            size=text_size,
+        )
+    if bus.ev_park is not None:
+        ax.plot(
+            x,
+            y,
+            marker=bus.ev_park.marker,
+            markeredgewidth=bus.ev_park.handle.get_markeredgewidth(),
+            markersize=bus.ev_park.size,
+            linestyle="None",
+            color=bus.color,
+            clip_on=False,
+            zorder=3,
+        )
+
 def _plot_line(
     ax: plt.axis,
     line: Line,
@@ -166,7 +245,7 @@ def _plot_line(
     text_size: int = 8,
 ):
     """
-    Plot lines
+    Plot line
 
     Parameters
     ----------
@@ -366,7 +445,7 @@ def _plot_intelligent_switch(
 
 def _plot_sensor(
     ax: plt.axis,
-    line: Line,
+    sensor: Sensor,
     text: bool = False,
     text_size: int = 8,
 ):
[…]
