+++ docs/source/conf.py
@@ -107,6 +107,11 @@ todo_include_todos = False
 #html_theme = "alabaster"sphinx_rtd_theme
 html_theme = "sphinx_book_theme"
 
+html_theme_options = {
+  "repository_url": "https://github.com/stinefm/relsad",
+  "use_repository_button": True,
+}
+
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
