+++ pyproject.toml
@@ -36,7 +36,6 @@ version_variable = [
     "pyproject.toml:version",
     "relsad/__init__.py:__version__",
 ]                                           # version location
-version_source = "tag"
 branch = "main"                             # branch to make releases of
 changelog_file = "CHANGELOG.rst"            # changelog file
 build_command = "poetry build"              # build dists
