### 2022-09-10 stinefm/relsad#1 @blthayer
README Suggestions
As part of JOSS review: https://github.com/openjournals/joss-reviews/issues/4516

- [x] "ICT" acronym used without definition, please spell out. I presume it's short for "information and communication technology?"
- [ ] (Optional) Style recommendation: shortening up the first paragraph to have a higher-level summary and moving some of the details into the "Features" section might allow the reader to more quickly comprehend the purpose of the tool
- [x] (Optional) In "Dependencies", you might consider linking to `pyproject.toml` for the single source of truth for dependencies
- [x] Markdown link in "Contributors" section isn't rendering correctly
- [ ] (Optional) A quick and simple code example in the README goes a long way. However, sometimes it's tough to make a quick and simple example for complex software :smile:
