### 2022-09-10 stinefm/relsad#2 @blthayer
Documentation Suggestions
As part of JOSS review: https://github.com/openjournals/joss-reviews/issues/4516

Please consider all suggestions below optional for the purposes of the JOSS review unless explicitly stated otherwise :smile:

- [x] On [home/landing page](https://relsad.readthedocs.io/en/latest/index.html), consider putting in-line links to pages in the second paragraph rather than simply italicizing the titles. *E.g.*, instead of "The _installation_-page", consider "The [Installation page](https://relsad.readthedocs.io/en/latest/installation.html)"
- [x] On [home/landing page](https://relsad.readthedocs.io/en/latest/index.html), consider adding a paragraph giving an overview of `RESLAD`. This could be sourced from the repo's [README](https://github.com/stinefm/relsad/blob/main/README.rst)
- On the [Installation page](https://relsad.readthedocs.io/en/latest/installation.html) I recommend updating the "Stable release" directions to use `python -m pip install ...` instead of just `pip install`. This is a gentle nudge for users to use virtual environments, and can also prevent issues/confusion. One situation I've seen and run into myself is if multiple versions of `pip` are on the PATH, the package gets installed into the "wrong" Python environment leading to confusion. Explicitly using `python -m pip ...` prevents this issue.
- [x] Consider pointing users to [pyenv](https://github.com/pyenv/pyenv) for Python management. It's an _excellent_ tool for installing Python and avoiding the pitfalls of using system package managers for installing/managing Python. I couldn't recommend it enough. It also makes it very simple and easy to run the latest version :smile:
- On the [Installation page](https://relsad.readthedocs.io/en/latest/installation.html), consider informing users what versions of Python are supported. Looking at `pyproject.toml`, it seems you support `3.8`, `3.9`, and `3.10`
- [x] (This suggestion's less optional :wink:) On the [Installation page](https://relsad.readthedocs.io/en/latest/installation.html) in the "From sources" section, the line `pip install poetry` is problematic. This is _not_ the [recommended way of installing poetry](https://python-poetry.org/docs/#installation). Using `pip` to install `poetry` is documented as the "advanced" way of installing `poetry`, and in previous versions of their documentation they explicitly recommended against it. I recommend pointing the reader to the `poetry` installation docs and changing the terminal directions to simply be:
$ cd relsad
$ poetry install
```
- [x] Consider using [sphinx-copybutton](https://sphinx-copybutton.readthedocs.io/en/latest/) so that tutorial code blocks can be copied with a single click.
- [x] (**Not optional**) - The tutorial has a bad import of `CustomDiscreteParameters`, which does not seem to exist in the current version of `relsad`:
from relsad.StatDist import (
    StatDist,
    StatDistType,
    NormalParameters,
    CustomDiscreteParameters,
)
```
result:
```
ImportError: cannot import name 'CustomDiscreteParameters' from 'relsad.StatDist' ...
```
- [ ] Related to the import error in the tutorial (and more errors below...), I **strongly** recommend configuring your test suite to run all documentation examples. This way, you'll know immediately if an update broke an example in the documentation. You can use the `--doctest-glob` argument to `pytest` (with an appropriate pattern, *e.g.* `*.rst`). You'll need to tell `pytest` where your documentation files live, too, perhaps via the `testpaths` configuration. [Here's the pytest documentation for running doctests](https://docs.pytest.org/en/7.1.x/how-to/doctest.html).
- [x] Consider wrapping comments in the tutorial onto new lines. At present, you have to use a scroll 2-3x the width of the code block to read the comment, as pictured below. I recommended following [PEP-8](https://peps.python.org/pep-0008/)'s suggestion of a 72 character limit for comments.

- [x] (**Not optional**): Tutorial has bad constructor call for creating a `Line`, specifically, `L6`: `TypeError: Line.__init__() got an unexpected keyword argument 'is_backup'`
- [x] (**Not optional**): The image in the docs for "Visualize topology" does not match the output when going through the tutorial. I copy + pasted all code directly from the tutorial, and only modified what was necessary to get by the errors listed above. I get an output that looks like the following:
![Figure 1](https://user-images.githubusercontent.com/14652512/189497118-0034c815-8e44-4a67-8a7b-2d7cdc29f6c9.png)

Perhaps `L3` is supposed to go from `B2` to `B4` instead of from `B3` to `B4` when the lines are created?

- [x] (**Not optional**): Creating the generation unit in the tutorial fails:
# A generation unit:

P1 = Production(
    name="P1",
    bus=B3,
    )

# A battery:

B1 = Battery(
    name="B1",
    bus=B6,
    )
Exception                                 Traceback (most recent call last)
Cell In [16], line 3
      1 # A generation unit:
----> 3 P1 = Production(
      4     name="P1",
      5     bus=B3,
      6     )
      8 # A battery:
     10 B1 = Battery(
     11     name="B1",
     12     bus=B6,
     13     )

File ~/PycharmProjects/relsad-joss/venv/lib/python3.10/site-packages/relsad/network/components/Production.py:85, in Production.__init__(self, name, bus, pmax, qmax)
     83     raise Exception("Production unit must be connected to a Bus")
     84 if bus.parent_network is not None:
---> 85     raise Exception(
     86         "Production unit must be created before the bus is connected to a network"
     87     )
     88 if pmax < 0:
     89     raise Exception(
     90         "The maximum active power that can be produced must be positive"
     91     )

Exception: Production unit must be created before the bus is connected to a network
```
I was able to work around this by creating the `Production` and `Battery` instances _before_ creating the `Distribution` instance. Not sure why this is necessary.

- [x] (**Not optional**): Creating the `EVPark` fails:
EVPark(
    name="EV1",
    bus=B5,
    num_ev=5,
    v2g_flag=True,
)
```
```
Cell In [18], line 1
----> 1 EVPark(
      2     name="EV1",
      3     bus=B5,
      4     num_ev=5,
      5     v2g_flag=True,
      6 )

TypeError: EVPark.__init__() got an unexpected keyword argument 'num_ev'
```

I'm going to submit this issue for now so that I can move to filing a different one, and then I'll update this issue as I continue through the documentation.
