### 2022-09-10 stinefm/relsad#3 @blthayer
`pytest` Configuration Suggestions
As part of JOSS review: https://github.com/openjournals/joss-reviews/issues/4516

Please consider this a completely optional suggestion from the JOSS review perspective, just sharing my knowledge, thoughts, and suggestions :smile:

The repo has a [pytest.ini](https://github.com/stinefm/relsad/blob/main/pytest.ini) file. `pytest` supports using [`pyproject.toml` for configuration](https://docs.pytest.org/en/7.1.x/reference/customize.html#pyproject-toml). I really like putting as much configuration as possible into `pyproject.toml` to keep my repositories nice and clean.

Now, on the _contents_ of the `pytest.ini` file:
- [x] Having `minversion` certainly doesn't hurt anything, but it's rendered more or less pointless since you're using `poetry` _and_ you're committing `poetry.lock` to your repo (which is great - I highly recommend that as a best practice). If you have a minimum package requirement, the common/best-practice place to define that is `pyproject.toml`
- [x] Explicitly "white listing" all the `testpaths` makes maintenance harder and will likely eventually cause a test to accidentally not be run. Since you're already properly naming your tests (placed in a directory called `tests`, all files start with `test_`), `pytest` will automatically find all your tests :smile:. So I suggest totally removing this from the configuration file.
