### 2022-09-10 stinefm/relsad#2 @blthayer
I'm going to pause my review for now until the errors/issues in the tutorial have been worked out. As mentioned above, I recommend running all examples as part of the testing suite using `pytest`'s `doctest` capabilities: https://docs.pytest.org/en/7.1.x/how-to/doctest.html
