### 2022-09-10 #4516 @blthayer
Update: I've begun my review, but am running into issues in the tutorial. See https://github.com/stinefm/relsad/issues/2

Please accept my apologies for the serious delays on my end in getting to this review.
