### 2022-09-11 #4516 @stinefm
Hi @blthayer, @fraukewiese,

Thank you for the thorough feedback! I have been working on improvements based on the task lists you have provided. I have tried to respond to all the provided suggestions where I have ticked off the tasks I have addressed. Below is a summary of the issues.

**Configuration Suggestions:**
Thank you for the good suggestions, I totally agree that including everything in the pyproject.toml file is cleaner and easier to maintain. Integrating the pytest options in the pyproject.toml file and exploiting the test naming convention was very nice.

**Documentation Suggestions:**
I agree that adding in-line links looks better than italic fonts. I added the overview paragraph in the readme file to the landing page as suggested. The installation page is now updated with pointers to pyenv, as well as modifications to the install commands as suggested. I also linked to the poetry installation page. All imports and initialization errors in the tutorial code are now corrected. I am still working on how I should include the tutorial code in the test suite. This will be done. The topology image you get by following the tutorial should now be correct.

**README Suggestions:**
I added the definition of ICT and referred to pyproject.toml for information about dependencies. I will evaluate other options for presenting the overall purpose of the tool on the README page. I struggle to find good ways of providing simple and quick code examples to present on the README page.

Best
Stine
