### 2022-09-11 e597b29
docs: improved tutorial and readme page based on review
