+++ docs/source/usage/tutorial/traditional_power_system/system.rst
@@ -1,0 +1,389 @@
+"""""""
+Imports
+"""""""
+
+To create a power system the necessary imports need to be added. 
+
+We will make use of `numpy <https://numpy.org/doc/stable/index.html>`_
+in this tutorial::
+    
+    import numpy as np
+
+For importing components from `RELSAD`::
+
+    from relsad.network.components import (
+        Bus,
+        Line,
+        Disconnector,
+        CircuitBreaker,
+        ManualMainController,
+    )
+
+
+For importing systems and networks from `RELSAD`:: 
+
+    from relsad.network.systems import (
+        PowerSystem,
+        Transmission,
+        Distribution,
+    )
+
+To run time dependent simulations, the time utilities of `RELSAD`
+must be imported::
+
+    from relsad.Time import (
+        Time, 
+        TimeUnit,
+        TimeStamp,
+    )
+
+Adding statistical distribution for, for example, outage time of
+components, is done by using the statistical distribution utilities
+of `RELSAD`, which needs to be imported::
+
+    from relsad.StatDist import (
+        StatDist,
+        StatDistType,
+        NormalParameters,
+    )
+
+The statistical distribution utilities of `RELSAD` enables a variety of custom distributions, including normal and uniform distributions.
+
+To prioritize the bus loadings during outages, the user may define
+cost functions that can be related to chosen buses. To use this feature,
+the `CostFunction` class must be imported::
+
+    from relsad.load.bus import CostFunction
+
+To plot the network topology, we import the `plot_topology` function::
+
+    from relsad.visualization.plotting import plot_topology
+
+The `Simulation` class must be imported to be able to run simulations::
+
+    from relsad.simulation import Simulation
+
+Finally, we import the following packages to read and plot some results::
+
+    import os
+    import pandas as pd
+    import matplotlib.pyplot as plt
+
+
+"""""""""""""""""
+Create components
+"""""""""""""""""
+
+First, we create the components for our system.
+
+Creating buses::
+
+    # Failure rate and outage time of the transformer on the bus
+    # are not necessary to add, this can be added on each bus.
+    # Their default values are 0 and Time(0) respectively.
+
+    B1 = Bus(
+        name="B1", 
+        n_customers=0, 
+        coordinate=[0, 0],
+    )
+    B2 = Bus(
+        name="B2", 
+        n_customers=1, 
+        coordinate=[1, 0],
+    )
+
+    B3 = Bus(
+        name="B3", 
+        n_customers=1, 
+        coordinate=[2, 1],
+    )
+
+    B4 = Bus(
+        name="B4", 
+        n_customers=1, 
+        coordinate=[2, 0],
+    )
+    
+    B5 = Bus(
+        name="B5", 
+        n_customers=1, 
+        coordinate=[3, 0],
+    )
+    
+    B6 = Bus(
+        name="B6", 
+        n_customers=1, 
+        coordinate=[3, 1],
+    ) 
+
+Creating lines:: 
+
+    # Failure rate and outage time of the lines can be added to each line.
+    # The default value of the line failure rate is 0, while the default
+    # outage time is 0 (Uniform float distribution with max/min values of 0).
+
+    # For adding statistical distributions, in this case a
+    # truncated normal distribution: 
+
+    line_stat_repair_time_dist = StatDist(
+        stat_dist_type=StatDistType.TRUNCNORMAL,
+        parameters=NormalParameters(
+            loc=1.25,
+            scale=1,
+            min_val=0.5,
+            max_val=2,
+        ),
+    )
+
[…]
