+++ docs/source/usage/tutorial/main.rst
@@ -1,0 +1,18 @@
+.. _tutorial:
+
+========
+Tutorial
+========
+
+Here, we describe the workflow of `RELSAD` containing the necessary steps
+for running a reliability analysis. First, multiple ways of defining
+the network is presented. Second, an example of how to run an analysis is
+shown. Finally, we show a way of visualizing the results.
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+   traditional_power_system/monte_carlo_simulation
+   traditional_power_system/sequential_simulation
+   active_power_systems

[…]
