### 2022-09-11 dd63f42
build: improved setup based on review
