+++ pyproject.toml
@@ -26,6 +26,7 @@ black = "^20.8b1"
 flake8 = "^3.9.0"
 pre-commit = "^2.12.0"
 python-semantic-release = "^7.31.4"
+sphinx-copybutton = "^0.5.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
@@ -69,3 +70,9 @@ ignore = "E203, W503"
 [tool.isort]
 profile = "black"
 skip = ["relsad/network/systems/__init__.py"]
+
+[tool.pytest.ini_options]
+addopts = "-ra -q"
+testpaths = [
+    "tests",
+]
