+++ pytest.ini
@@ -1,36 +1,0 @@
-# pytest.ini
-[pytest]
-minversion = 6.0
-addopts = -ra -q
-testpaths = 
-    tests/test_battery.py
-    tests/test_bus.py
-    tests/test_line.py
-    tests/test_production.py
-    tests/test_circuitbreaker.py
-    tests/test_component_interaction.py
-    tests/test_disconnector.py
-    tests/test_evpark.py
-    tests/test_network_behavior_full_support_before_sectioning_time.py
-    tests/test_network_behavior_full_support_after_sectioning_time.py
-    tests/test_network_behavior_limited_support_before_sectioning_time.py
-    tests/test_network_behavior_limited_support_after_sectioning_time.py
-    tests/test_network_behavior_survival_before_sectioning_time.py
-    tests/test_network_behavior_survival_after_sectioning_time.py
-    tests/test_load_flow.py
-    tests/test_energy_shed.py
-    tests/test_Time.py
-    tests/test_TimeStamp.py
-    tests/test_Table.py
-    tests/test_sensor.py
-    tests/test_microgrid.py
-    tests/test_section.py
-    tests/test_powersystem.py
-    tests/test_subsystem.py
-    tests/test_simulation.py
-    tests/test_system_config.py
-    tests/test_topology_ICT_dfs.py
-    tests/test_utils.py
-    tests/test_rbts_bus5_feeder_1.py
-    tests/test_examples.py
-
