+++ relsad/examples/CINELDI/network.py
@@ -799,7 +799,7 @@ def initialize_network(
     )
     L3 = Line(
         name="L3",
-        fbus=B3,
+        fbus=B2,
         tbus=B4,
         r=0.7114,
         x=0.2351,
@@ -843,7 +843,7 @@ def initialize_network(
     DL1b = Disconnector("L1b", L1, B2)
     DL2a = Disconnector("L2a", L2, B2)
     DL2b = Disconnector("L2b", L2, B3)
-    DL3a = Disconnector("L3a", L3, B3)
+    DL3a = Disconnector("L3a", L3, B2)
     DL3b = Disconnector("L3b", L3, B4)
     DL4a = Disconnector("L4a", L4, B4)
     DL4b = Disconnector("L4b", L4, B5)
