### 2022-09-11 9a424d8
fix: fixed bug in controller
