+++ relsad/network/components/DistributionController.py
@@ -277,7 +277,11 @@ class DistributionController(Component, Controller):
         # Loop disconnected sections
         for section in disconnected_sections:
             sensors = unique(
-                [switch.line.sensor for switch in section.switches]
+                [
+                    switch.line.sensor
+                    for switch in section.switches
+                    if switch.line.sensor is not None
+                ]
             )
             num_fails = 0
             need_manual_attention = False
@@ -321,7 +325,11 @@ class DistributionController(Component, Controller):
         # Loop connected sections
         for section in connected_sections:
             sensors = unique(
-                [switch.line.sensor for switch in section.switches]
+                [
+                    switch.line.sensor
+                    for switch in section.switches
+                    if switch.line.sensor is not None
+                ]
             )
             num_fails = 0
             need_manual_attention = False
