### 2022-09-12 a04b4a3
docs: added sphinx-copybutton to dependencies
