+++ docs/requirements.txt
@@ -1,2 +1,3 @@
 sphinxcontrib-bibtex == 2.4.2
 sphinx-book-theme == 0.3.2
+sphinx-copybutton == 0.5.0
