### 2022-09-14 b6de186
docs: added code coverage badge
