+++ README.rst
@@ -1,8 +1,9 @@
+
+|language tab| |code coverage tab| |contributors tab| |python version tab| |relsad version tab| |build tab| |license tab|
+
 ######
 RELSAD
 ######
-
-|language tab| |contributors tab| |python version tab| |relsad version tab| |build tab| |license tab|
 
 .. overview_start
 
