### 2022-09-14 f4b3f77
fig: fixed bug in code coverage badge
