### 2022-09-15 d960cc8
docs: added joss badge
