### 2022-09-19 stinefm/relsad#1 @blthayer
I'll go ahead and close this out - the README is certainly looking better :smile:

As far as the code example - I wonder if it would be easy to whip up and demonstrate a simple two-bus, single-line, one-load, one-generator example?
