### 2022-09-20 stinefm/relsad#4 @blthayer
Documentation Suggestions, Round Two: "Usage" Section
As part of JOSS review: https://github.com/openjournals/joss-reviews/issues/4516

Please consider all suggestions below optional for the purposes of the JOSS review unless explicitly stated otherwise :smile:

For this issue/ticket, I'm focusing on the tutorials (the "Usage" section). Later in the week, I'll read through the "Background" and "Input Data" sections and provide feedback.

- [ ] Typo in the [Monte Carlo tutorial](https://relsad.readthedocs.io/en/latest/usage/tutorial/traditional_power_system/monte_carlo_simulation.html#traditional-power-system-monte-carlo) -> "Mont Carlo" vs. "Monte Carlo"
- [ ] If the imports, system under study, _etc._ are the same for both the "Monte Carlo" and "Sequential" tutorials, I would recommend creating a section named something like "Set Up Simulation" which contains all the imports, creation of lines, buses, _etc._ and then make the "Monte Carlo" and "Sequential Simulation" sections focused on setting up the different simulations and viewing the results. This will take a lot of repetition out, which is good for both readers and maintainers :smile:
- [ ] (**Not Optional**): When I run the [sequential simulation](https://relsad.readthedocs.io/en/latest/usage/tutorial/traditional_power_system/sequential_simulation.html) tutorial, I get a "flat line" result, unlike in the tutorial:
- [ ] (**Not Optional**): There's an error in the [Active power systems](https://relsad.readthedocs.io/en/latest/usage/tutorial/active_power_systems.html) tutorial. When attempting to create the variable `P1 = Production(name="P1", bus=B3)`, the following exception is raised: `Exception: Production unit must be created before the bus is connected to a network`. You can actually see a similar (but somewhat confusingly, different) error in the tutorial itself: "NB! The active generation units must be initialized before any network initializations." I've worked around this for now in order to continue my review by moving the instantiation of the `PowerSystem`, `Transmission`, and `Distribution` objects until _after_ initializing the `Production` object.
- [ ] (**Not Optional**): There's an error in the [Power system with ideal ICT network](https://relsad.readthedocs.io/en/latest/usage/tutorial/active_power_systems.html#power-system-with-ideal-ict-network) tutorial. When creating the sensor object via `S1 = Sensor(name="S1", line=L2)`, the following exception is raised: `Exception: Sensor must be created before the line is connected to a network`. It's not clear to me if the line should be "ML2" instead of "L2", but even with that change I get the same error. For now, I've worked around this issue by moving the instantiation of `S1` to prior to the instantiation of the `Microgrid` and `Distribution` objects.
- [ ] (**Not Optional**): There are a couple typos in the [Power system with failable ICT network](https://relsad.readthedocs.io/en/latest/usage/tutorial/active_power_systems.html#power-system-with-failable-ict-network) tutorial that lead to errors. Specifically, the lines `ICTL1` and `ICTL3` refer to an object `ICTNISWS1` which does not exist. It appears these objects should reference `ICTNISW1` instead.
- [ ] Oddly enough, in the majority of English dictionaries (as far as I can tell), "failable" is not a word - you _might_ looking for the word "[fallable](https://www.merriam-webster.com/dictionary/fallible)"
- [ ] The [CINELDI example](https://relsad.readthedocs.io/en/latest/usage/CINELDI_example.html) seems to assume the user has checked out the repository from source rather than installing from PyPI, but doesn't specify as much. Mentioning the source code is required and providing a link back to the GitHub repo could be useful for readers.
