### 2022-09-20 stinefm/relsad#2 @blthayer
@stinefm - I'm going to go ahead and close this ticket out since you've addressed 11/12 suggestions. Things are looking much improved, nice work!

I will go ahead and file a separate issue for using `pytest` for the documentation. That is _not_ required for the JOSS review, but I think will save you, other maintainers, and users a lot of time in the long run by being 100% certain that the tutorials work when code is changed.
