### 2022-09-20 #4516 @blthayer
> @blthayer : Are the issues you mentioned in the tutorial solved from your point of view?

@fraukewiese @stinefm

Stine has done an _excellent_ job addressing my first round of feedback. I have found some additional errors/issues and filed them [here](https://github.com/stinefm/relsad/issues/4).

I plan to continue my review later this week, but no sooner than Wednesday evening after work. In the worst case, I'll be picking this back up this upcoming weekend. Again - I'm sincerely sorry for the delays on my end, and I'll be trying to find time wherever I can so that we can get this over the finish line.
