### 2022-09-20 bd4b509
docs: fixed typo in README
