### 2022-09-20 addeba2
build: removed pytest options from pyproject
