+++ pyproject.toml
@@ -72,7 +72,6 @@ profile = "black"
 skip = ["relsad/network/systems/__init__.py"]
 
 [tool.pytest.ini_options]
-addopts = "-ra -q"
 testpaths = [
     "tests",
 ]
