### 2022-09-20 be32cb7
docs: restructured, ported code to relsad/examples
