### 2022-09-20 7880664
tests: added tests for docs code
