+++ tests/test_docs_tutorial_power_system_with_fallible_ict.py
@@ -1,0 +1,6 @@
+import runpy
+
+def test_tutorial_power_system_with_fallible_ict():
+    runpy.run_module(
+        "relsad.examples.tutorial.power_system_with_fallible_ict"
+    )
