+++ tests/test_docs_tutorial_islanded_networks.py
@@ -1,0 +1,6 @@
+import runpy
+
+def test_tutorial_islanded_networks():
+    runpy.run_module(
+        "relsad.examples.tutorial.islanded_networks"
+    )
