+++ tests/test_docs_tutorial_grid_connected_microgrids.py
@@ -1,0 +1,6 @@
+import runpy
+
+def test_tutorial_grid_connected_microgrids():
+    runpy.run_module(
+        "relsad.examples.tutorial.grid_connected_microgrids"
+    )
