+++ tests/test_docs_tutorial_system.py
@@ -1,0 +1,6 @@
+import runpy
+
+def test_tutorial_system():
+    runpy.run_module(
+        "relsad.examples.tutorial.system"
+    )
