+++ tests/test_docs_tutorial_active_generation_units.py
@@ -1,0 +1,6 @@
+import runpy
+
+def test_tutorial_active_generation_units():
+    runpy.run_module(
+        "relsad.examples.tutorial.active_generation_units"
+    )
