### 2022-09-20 #4516 @stinefm
Hi @blthayer, @fraukewiese,

Thank you for the feedback! I have worked on the issues you addressed. Below is a summary of the issues.

Documentation Suggestions:
The typos and spellings you mentioned are fixed. I have created a "Setup" section for the "Traditional power system" tutorial as suggested to improve maintainability and readability. The "Sequential simulation" tutorial contained a bug which is fixed now. The issues encountered in the "Active power systems" section is due to the fact that the components must be imported before networks are initiated. I have moved up the sentences specifying this. The source code is now mentioned as the location of the "CINELDI example".

Test suite for tutorial code:
The code in the tutorials are now integrated in the test suite. I was not able to use doctest as suggested since it is only compatible with single command tests. The script based tutorials are now included in the source code where they are executed using runpy. I totally agree that integrating the tutorial code in the test suite is the way to go!

Best
Stine
