### 2022-09-20 a3ace6c
docs: fixed usage link in README
