+++ docs/source/background/Components/dg.rst
@@ -6,11 +6,10 @@ Load and Generation unit implementation
 Implementation of load
 .....................................
 
-In RELSAD, a load can be set on each bus. The load can either be set as a constant load or a time varied load. For time varied loads, the load can be allocated to the bus through an array containing the load for each time increment (see :meth:`Bus.add_load_data() <relsad.network.components.Bus.Bus.add_load_data>`). The package allows user-defined time increments and the user can set a start and an end time. If the load array varies from the user-specified time increment, the load data will be interpolated to fit the wanted time increment. 
+In RELSAD, a load can be set on each bus. The load can either be set as a constant load or a time-varied load. For time-varied loads, the load can be allocated to the bus through an array containing the load for each time increment (see :meth:`Bus.add_load_data() <relsad.network.components.Bus.Bus.add_load_data>`). The package allows user-defined time increments and the user can set a start and an end time. If the load array varies from the user-specified time increment, the load data will be interpolated to fit the wanted time increment. 
 
-RELSAD supports the inclusion of multiple load types and the specific interruption cost related to the different customer categories. The specific inturrption cost can be added in order to calculate the Cost of Energy Not Supplied in the system (default value for the specific interruption cost is set to 1). 
+RELSAD supports the inclusion of multiple load types and the specific interruption cost related to the different customer categories. The specific interruption cost can be added in order to calculate the Cost of Energy Not Supplied in the system (the default value for the specific interruption cost is set to 1). 
 In addition, multiple different load types can be added to a bus, making the bus able to have different customer categories connected. 
-
 .. 
     the specific interruption cost for each customer category can be included for calculating the Cost of Energy Not Supplied in the system. 
 
