### 2022-09-21 dba9f76
refactor: changed default argument and ict node coordinates
