### 2022-09-21 274eda2
refactor: removed unused variables
