### 2022-09-21 bec6be9
refactor: modified ICTLine fail behavior
