### 2022-09-27 stinefm/relsad#4 @blthayer
Thanks for addressing these, @stinefm!

One more recommendation: as part of the [Active power systems](https://relsad.readthedocs.io/en/latest/usage/tutorial/active_power_systems.html) tutorial, I recommend emphasizing (probably with bold? Up to you) the following sentence:

> The active generation units must be initialized before any network initializations.

Additionally, it wouldn't hurt to clarify a bit for new users. What if the sentence read:

"The active generation units must be initialized before any network initializations, *e.g.*, before `PowerSystem`, `Transmission`, or `Distribution` objects are initialized."
