### 2022-09-27 stinefm/relsad#6 @blthayer
Suggestions for "Background" section of docs
As part of JOSS review: https://github.com/openjournals/joss-reviews/issues/4516

Please consider all suggestions optional for JOSS review unless otherwise specified.

- [x] (**Not optional**) [Battery behavior](https://relsad.readthedocs.io/en/latest/background/Components/battery.html): E_bat_charge and E_bat_discharge are incorrect in that power and energy are used interchangeably without multiplying power by time to get energy. Suggest adding "delta t" terms
- [x] [Implementation of load](https://relsad.readthedocs.io/en/latest/background/Components/dg.html#implementation-of-load) "...the load data will be interpolated..." - I recommend specifying the method of interpolation. Simple linear interpolation? (nothing wrong with linear interpolation, but I think it's worth mentioning to the user)
- [x] [Implementation of load](https://relsad.readthedocs.io/en/latest/background/Components/dg.html#implementation-of-load): formatting issue

- [x] [Implementation of generation units](https://relsad.readthedocs.io/en/latest/background/Components/dg.html#implementation-of-generation-units): I suggest changing "are implemented generic," with "are implemented generically" for correct grammar
- [x] [Implementation ... and vehicle to grid](https://relsad.readthedocs.io/en/latest/background/Components/ev.html#implementation-of-electric-vehicles-and-vehicle-to-grid): The first sentence is confusing, particularly the "with drawn SOC level." Is it supposed to say "with _randomly_ drawn SOC level?" If so, I would also recommend specifying which distribution is used: uniform, Gaussian, other?
- [x] [Backward-Forward sweep](https://relsad.readthedocs.io/en/latest/background/Loadflow/bfs.html#backward-forward-sweep): I suggest using a different citation for discussing convergence issues from Newton-Raphson. The earliest citation that I'm aware of is:
- Title: **Cases of difficult convergence in load-flow problems**
- Author: C. Trevino
- IEEE Paper no. 71-62-PWR
- Presented at the _IEEE Summer Power Meeting_, Los Angeles, CA, 1970
(this is the paper cited by the textbook **Distribution System Modeling and Analysis**, 3rd ed, by William H. Kersting)

- [ ] [Backward-Forward sweep](https://relsad.readthedocs.io/en/latest/background/Loadflow/bfs.html#backward-forward-sweep): It seems like a citation is in order for the Backward-Forward sweep algorithm/mathematics. My understanding is that William Kersting et al pioneered what he called the "Ladder Iterative Technique" (now commonly known as forward-backward sweep or backward-forward sweep depending on how you start the algorithm) back in the 1970s. The citation provided in his book is as follows:

> Kersting, W.H. and Mendive, D.L., An application of ladder network theory to the solution of three-phase radial load-flow problems, IEEE Conference Paper, Presented at the IEEE Winter Power Meeting, New York, January 1976

- [x] [Backward-Forward sweep](https://relsad.readthedocs.io/en/latest/background/Loadflow/bfs.html#backward-forward-sweep): Typo: "can be calculated battery" - shows up in two places
- [x] (**Not optional**): [Forward sweep](https://relsad.readthedocs.io/en/latest/background/Loadflow/bfs.html#forward-sweep): The documentation is not rendering for T1, T2, or voltage angle expression:

- [x] [Load shedding optimization problem](https://relsad.readthedocs.io/en/latest/background/Loadflow/opt.html#load-shedding-optimization-problem): Suggest changing "minimize the total shedded load" to "minimize the total load shed" (English is a bizarre language :smile:). The next paragraph also uses "shedded," which is not technically a word
- [ ] [Markov models](https://relsad.readthedocs.io/en/latest/background/Reliability/markov.html#markov-models) I recommend adding a citation for the reader to dig into Markov models
- [x] [Markov models](https://relsad.readthedocs.io/en/latest/background/Reliability/markov.html#markov-models): "bad" reference: "A simple three-state model as seen in Figure REF is used"
