### 2022-09-27 #4516 @blthayer
@stinefm - as I got further into the documentation I found that you're using forward-backward sweep (FBS) (or backward-forward sweep, same thing :smile:) for solving the power flow. My understanding is that FBS is only valid for radial systems. With that in mind, I have two suggestions:
- Mention up front (first or second paragraph in README page and documentation) that RELSAD is _only_ for radial systems
- If you haven't already (you might have, I haven't double-checked), ensure that the an exception is raised if a user creates a meshed system.

A tool I've used in the past for graph-theory work/analysis in Python is `networkx`. I haven't looked under the covers at the RELSAD source code, but it's possible that building up a `networkx` graph for the system could be an effective way of catching loops before solving the power flow.
