### 2022-09-28 13c416b
test: added test for is_cyclic
