### 2022-09-28 390b833
docs: fixes based on review from blthayer
