+++ docs/source/conf.py
@@ -100,6 +100,7 @@ pygments_style = "sphinx"
 # If true, `todo` and `todoList` produce output, else they produce nothing.
 todo_include_todos = False
 
+numfig = True
 
 # -- Options for HTML output -------------------------------------------
 
