+++ docs/source/references.bib
@@ -10,6 +10,13 @@
   publisher={Elsevier}
 }
 
+@inproceedings{trevino1970cases,
+  title={Cases of difficult convergence in load-flow problems},
+  author={Trevino, C},
+  booktitle={IEEE Summer Power Meeting},
+  year={1970}
+}
+
 @article{jones2002modelling,
   title={A modelling method for building-integrated photovoltaic power supply},
   author={Jones, AD and Underwood, CP},

[…]
