### 2022-09-28 29194d3
refactor: cleaned up TEST10 example network code
