### 2022-09-28 0dd7d22
refactor: added check for radial network during bfs loadflow
