+++ docs/source/references.bib
@@ -47,3 +47,10 @@
   	Year = 1992
 }
 
+@book{ccinlar2011probability,
+  title={Probability and stochastics},
+  author={{\c{C}}inlar, Erhan and E. {\dh}C{\i}nlar},
+  volume={261},
+  year={2011},
+  publisher={Springer}
+}
