### 2022-10-09 #4516 @blthayer
@fraukewiese - my review is complete. Please let me know if anything further is needed from me. Thank you!

@stinefm - I hope you consider volunteering to review for JOSS. There aren't too many in the power systems engineering field on the reviewer list/spreadsheet.
