### 2022-10-12 fcbf181
build: updated semantic versioning setup
