+++ CHANGELOG.rst
@@ -3,8 +3,3 @@ Changelog
 =========
 
 <!--next-version-placeholder-->
-
-v0.1.0 (30/08/2022)
--------------------
-
-- First release of `relsad`
