+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "relsad"
-version = "0.1.0"
+version = "0.2.11"
 description = "A package that facilitates reliability investigations of power systems"
 readme = "README.rst"
 repository = "https://github.com/stinefm/relsad"
@@ -37,7 +37,6 @@ version_variable = [
     "pyproject.toml:version",
     "relsad/__init__.py:__version__",
 ]                                           # version location
-version_source = "tag"                      # version source
 branch = "main"                             # branch to make releases of
 changelog_file = "CHANGELOG.rst"            # changelog file
 build_command = "poetry build"              # build dists
