### 2022-10-12 6198df3
0.2.12
Automatically generated by python-semantic-release
