+++ CHANGELOG.rst
@@ -3,3 +3,7 @@ Changelog
 =========
 
 <!--next-version-placeholder-->
+
+## v0.2.12 (2022-10-12)
+### Documentation
+* Fixed formatting ([`ac21263`](https://github.com/stinefm/relsad/commit/ac21263f527f1c571ea636224585c625e76d2cea))
