### 2022-10-12 97b94ac
build: changed CHANGELOG file type
