+++ CHANGELOG.md
@@ -1,6 +1,4 @@
-=========
-Changelog
-=========
+# Changelog
 
 <!--next-version-placeholder-->
 
