### 2022-10-13 #4516 @fraukewiese
@stinefm : Could you please check the following in the paper:

Line 50 - can the brackets around DSO be deleted?
Line 96 - I think the comma behind Figure 1 can be deleted
Line 120 - this is the only reference where you do not display the https://doi.org/ before the number. Would be good to make it consistent the way the other references are displayed. And it directs to https://10.0.3.239/978-1-4615-7728-7 , so I think also here, the doi.org/ is issing - Please check
