### 2022-10-13 #4516 @fraukewiese
@stinefm : And I agree to @blthayer that it would be great if you would volunteer to review for JOSS! :)
