+++ joss/paper.bib
@@ -1,5 +1,6 @@
 @book{Billinton:1992,
-  	url = {10.1007/978-1-4615-7728-7},
+  	doi = {10.1007/978-1-4615-7728-7},
+  	url = {https://link.springer.com/book/10.1007/978-1-4615-7728-7},
   	Author = {{Billinton}, R. and {Allan}, R. N.},
   	Booktitle = {Reliability evaluation of engineering systems: First Edition, by Roy Billinton and Ronald N. Allan.~ISBN 978-1-4615-7728-7 (eBook).~Published by Springer, Springer, Boston, MA, 1992.},
   	Publisher = {Springer},
@@ -30,8 +31,6 @@
   year={2018},
   publisher={Elsevier}
 }
-
-<!--- Kanskje ---> 
 
 @article{celli:2013,
   title={Reliability assessment in smart distribution networks},
