+++ joss/paper.md
@@ -29,10 +29,6 @@ affiliations:
 date: 19 March 2022
 bibliography: paper.bib
 
-# Optional fields if submitting to a AAS journal too, see this blog post:
-# https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
-#aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
-#aas-journal: Astrophysical Journal <- The name of the AAS journal.
 ---
 
 # Summary
@@ -42,12 +38,7 @@ The electrical power distribution system is under constant development and multi
 The traditional reliability analysis methods for power systems do not consider the new components and technology, and the behavior and impact these have on the distribution network. Therefore, new considerations need to be taken to address these new changes in the system. 
 This paper presents an open-source reliability assessment tool for smart and active distribution systems named `RELSAD`. 
 
-`RELSAD` -- RELiability tool for Smart and Active Distribution networks is a Python-based reliability assessment tool that aims to function as a foundation for reliability calculation of modern distribution systems. RELSAD can be used by scientists, engineers, and Distribution System Operators (DSO) for the planning and operation of modern distribution system where the goal is to investigate the reliability of a given system.
-<!--`RELSAD` is a Python-based package tool that aims to function as a foundation for reliability calculation for smart and active distribution systems. -->
-The tool allows for Monte Carlo simulation based reliability analysis of modern distribution networks, and sequential simulation of the network behavior with user-defined loading and failure evolution to investigate the impact of the introduction of for instance ICT components. The package supports user-selected time increment steps over a user-defined time period. In the tool, active components such as microgrids, distributed generation, batteries, and electrical vehicles are implemented. To evaluate smart power systems, ICT components such as automated switches, sensors, and control system for the power grid are also implemented. In addition to component implementation, in order to evaluate the reliability of such complex systems, the complexity, dependencies within a system, and interdependencies between systems and components are accounted for. 
-
-
-<!--- the dependencies between the different components and networks are included. --->
+`RELSAD` -- RELiability tool for Smart and Active Distribution networks is a Python-based reliability assessment tool that aims to function as a foundation for reliability calculation of modern distribution systems. RELSAD can be used by scientists, engineers, and Distribution System Operators (DSO) for the planning and operation of modern distribution system where the goal is to investigate the reliability of a given system. The tool allows for Monte Carlo simulation based reliability analysis of modern distribution networks, and sequential simulation of the network behavior with user-defined loading and failure evolution to investigate the impact of the introduction of for instance ICT components. The package supports user-selected time increment steps over a user-defined time period. In the tool, active components such as microgrids, distributed generation, batteries, and electrical vehicles are implemented. To evaluate smart power systems, ICT components such as automated switches, sensors, and control system for the power grid are also implemented. In addition to component implementation, in order to evaluate the reliability of such complex systems, the complexity, dependencies within a system, and interdependencies between systems and components are accounted for. 
 
 `RELSAD` offers the following features:
 
@@ -55,27 +46,15 @@ The tool allows for Monte Carlo simulation based reliability analysis of modern 
 
 * An extensive reliability analysis foundation with important and new reliability indices that aims to give a complete picture of the reliability of both the power system and the components in the system.
 
-* Inclusion of active participation of different power sources such as distributed generation, microgrids and batteries. <!--A simulation tool that includes active participation of different power sources such as distributed generation, microgrids, and batteries. -->
+* Inclusion of active participation of different power sources such as distributed generation, microgrids and batteries.
 
[…]
