### 2022-10-13 #4516 @stinefm
@fraukewiese, I have now solved the minor issues regarding the paper.
I have followed the checklist and archived the reviewed software in Zenodo.
The DOI is: 10.5281/zenodo.7195034
The version tag of the archived version is: 0.2.14.

I want to thank you all @fraukewiese, @robinroche, and @blthayer for the nice reviewing process!
I will make sure to volunteer after my PhD is delivered.

Best
Stine
